Base.Markdown.Header{1}

Base.Cartesian

Base.Markdown.Paragraph

The (non-exported) Cartesian module provides macros that facilitate writing multidimensional algorithms. It is hoped that Cartesian will not, in the long term, be necessary; however, at present it is one of the few ways to write compact and performant multidimensional code.

Base.Markdown.Header{2}

Principles of usage

Base.Markdown.Paragraph

A simple example of usage is:

Base.Markdown.Code

@nloops 3 i A begin
    s += @nref 3 A i
end

Base.Markdown.Paragraph

which generates the following code:

Base.Markdown.Code

for i_3 = 1:size(A,3)
    for i_2 = 1:size(A,2)
        for i_1 = 1:size(A,1)
            s += A[i_1,i_2,i_3]
        end
    end
end

Base.Markdown.Paragraph

In general, Cartesian allows you to write generic code that contains repetitive elements, like the nested loops in this example. Other applications include repeated expressions (e.g., loop unwinding) or creating function calls with variable numbers of arguments without using the "splat" construct (i...).

Base.Markdown.Header{2}

Basic syntax

Base.Markdown.Paragraph

The (basic) syntax of @nloops is as follows:

Base.Markdown.List

The first argument must be an integer (not a variable) specifying the number of loops.

The second argument is the symbol-prefix used for the iterator variable. Here we used i, and variables i_1, i_2, i_3 were generated.

The third argument specifies the range for each iterator variable. If you use a variable (symbol) here, it's taken as 1:size(A,dim). More flexibly, you can use the anonymous-function expression syntax described below.

The last argument is the body of the loop. Here, that's what appears between the begin...end.

Base.Markdown.Paragraph

There are some additional features of @nloops described in the reference section.

Base.Markdown.Paragraph

@nref follows a similar pattern, generating A[i_1,i_2,i_3] from @nref 3 A i. The general practice is to read from left to right, which is why @nloops is @nloops 3 i A expr (as in for i_2 = 1:size(A,2), where i_2 is to the left and the range is to the right) whereas @nref is @nref 3 A i (as in A[i_1,i_2,i_3], where the array comes first).

Base.Markdown.Paragraph

If you're developing code with Cartesian, you may find that debugging is easier when you examine the generated code, using macroexpand:

Base.Markdown.Code

julia> macroexpand(:(@nref 2 A i))
:(A[i_1,i_2])

Base.Markdown.Header{3}

Supplying the number of expressions

Base.Markdown.Paragraph

The first argument to both of these macros is the number of expressions, which must be an integer. When you're writing a function that you intend to work in multiple dimensions, this may not be something you want to hard-code. If you're writing code that you need to work with older Julia versions, currently you should use the @ngenerate macro described in an older version of this documentation.

Base.Markdown.Paragraph

Starting in Julia 0.4-pre, the recommended approach is to use a @generated function. Here's an example:

Base.Markdown.Code

@generated function mysum{T,N}(A::Array{T,N})
    quote
        s = zero(T)
        @nloops $N i A begin
            s += @nref $N A i
        end
        s
    end
end

Base.Markdown.Paragraph

Naturally, you can also prepare expressions or perform calculations before the quote block.

Base.Markdown.Header{3}

Anonymous-function expressions as macro arguments

Base.Markdown.Paragraph

Perhaps the single most powerful feature in Cartesian is the ability to supply anonymous-function expressions that get evaluated at parsing time. Let's consider a simple example:

Base.Markdown.Code

@nexprs 2 j->(i_j = 1)

Base.Markdown.Paragraph

@nexprs generates n expressions that follow a pattern. This code would generate the following statements:

Base.Markdown.Code

i_1 = 1
i_2 = 1

Base.Markdown.Paragraph

In each generated statement, an "isolated" j (the variable of the anonymous function) gets replaced by values in the range 1:2. Generally speaking, Cartesian employs a LaTeX-like syntax. This allows you to do math on the index j. Here's an example computing the strides of an array:

Base.Markdown.Code

s_1 = 1
@nexprs 3 j->(s_{j+1} = s_j * size(A, j))

Base.Markdown.Paragraph

would generate expressions

Base.Markdown.Code

s_1 = 1
s_2 = s_1 * size(A, 1)
s_3 = s_2 * size(A, 2)
s_4 = s_3 * size(A, 3)

Base.Markdown.Paragraph

Anonymous-function expressions have many uses in practice.

Base.Markdown.Header{4}

Macro reference

# Base.Cartesian.@nloops — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Cartesian.@nref — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Cartesian.@nextract — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Cartesian.@nexprs — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Cartesian.@ncall — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Cartesian.@ntuple — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Cartesian.@nall — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Cartesian.@nany — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Cartesian.@nif — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.