Base.Markdown.Header{1}

Documentation

Base.Markdown.Paragraph

Julia enables package developers and users to document functions, types and other objects easily via a built-in documentation system since Julia 0.4.

Base.Markdown.Admonition

This documentation system can also be used in Julia 0.3 via the Docile.jl package; see the documentation for that package for more details.

Base.Markdown.Paragraph

The basic syntax is very simple: any string appearing at the top-level right before an object (function, macro, type or instance) will be interpreted as documenting it (these are called docstrings). Here is a very simple example:

Base.Markdown.Code

"Tell whether there are too foo items in the array."
foo(xs::Array) = ...

Base.Markdown.Paragraph

Documentation is interpreted as Markdown, so you can use indentation and code fences to delimit code examples from text. Technically, any object can be associated with any other as metadata; Markdown happens to be the default, but one can construct other string macros and pass them to the @doc macro just as well.

Base.Markdown.Paragraph

Here is a more complex example, still using Markdown:

Base.Markdown.Code

"""
    bar(x[, y])

Compute the Bar index between `x` and `y`. If `y` is missing, compute
the Bar index between all pairs of columns of `x`.

# Examples
```julia
julia> bar([1, 2], [1, 2])
1
```
"""
function bar(x, y) ...

Base.Markdown.Paragraph

As in the example above, we recommend following some simple conventions when writing documentation:

Base.Markdown.List

Always show the signature of a function at the top of the documentation, with a four-space indent so that it is printed as Julia code.

Include a single one-line sentence describing what the function does or what the object represents after the simplified signature block. If needed, provide more details in a second paragraph, after a blank line.

Do not repeat yourself.

Only provide an argument list when really necessary.

Base.Markdown.Paragraph

For simple functions, it is often clearer to mention the role of the arguments directly in the description of the function's purpose. An argument list would only repeat information already provided elsewhere. However, providing an argument list can be a good idea for complex functions with many arguments (in particular keyword arguments). In that case, insert it after the general description of the function, under an # Arguments header, with one * bullet for each argument. The list should mention the types and default values (if any) of the arguments:

Base.Markdown.Code

"""
...
# Arguments
* `n::Integer`: the number of elements to compute.
* `dim::Integer=1`: the dimensions along which to perform the computation.
...
"""

Base.Markdown.List

Group examples under an # Examples section and use ``julia blocks instead of standard text.

Use backticks to identify code and equations.

Place the starting and ending """ characters on lines by themselves.

Base.Markdown.Paragraph

That is, write:

Base.Markdown.Code

"""
...

...
"""
f(x, y) = ...

Base.Markdown.Paragraph

rather than:

Base.Markdown.Code

"""...

..."""
f(x, y) = ...

Base.Markdown.Paragraph

This makes it more clear where docstrings start and end.

Base.Markdown.List

Respect the line length limit used in the surrounding code.

Base.Markdown.Header{2}

Accessing Documentation

Base.Markdown.Paragraph

Documentation can be accessed at the REPL or in IJulia by typing ? followed by the name of a function or macro, and pressing Enter. For example,

Base.Markdown.Code

?fft
?@time
?r""

Base.Markdown.Paragraph

will bring up docs for the relevant function, macro or string macro respectively. In Juno using Ctrl-D will bring up documentation for the object under the cursor.

Base.Markdown.Header{2}

Functions & Methods

Base.Markdown.Paragraph

Functions in Julia may have multiple implementations, known as methods. While it's good practice for generic functions to have a single purpose, Julia allows methods to be documented individually if necessary. In general, only the most generic method should be documented, or even the function itself (i.e. the object created without any methods by function bar end). Specific methods should only be documented if their behaviour differs from the more generic ones. In any case, they should not repeat the information provided elsewhere. For example:

Base.Markdown.Code

"""
Multiplication operator. `x*y*z*...` calls this function with multiple
arguments, i.e. `*(x,y,z...)`.
"""
function *(x, y)
  # ... [implementation sold separately] ...
end

"When applied to strings, concatenates them."
function *(x::AbstractString, y::AbstractString)
  # ... [insert secret sauce here] ...
end

help?>*
Multiplication operator. `x*y*z*...` calls this function with multiple
arguments, i.e. `*(x,y,z...)`.

When applied to strings, concatenates them.

Base.Markdown.Paragraph

When retrieving documentation for a generic function, the metadata for each method is concatenated with the catdoc function, which can of course be overridden for custom types.

Base.Markdown.Header{2}

Advanced Usage

Base.Markdown.Paragraph

The @doc macro associates its first argument with its second in a per-module dictionary called META. By default, documentation is expected to be written in Markdown, and the doc"" string macro simply creates an object representing the Markdown content. In the future it is likely to do more advanced things such as allowing for relative image or link paths.

Base.Markdown.Paragraph

When used for retrieving documentation, the @doc macro (or equally, the doc function) will search all META dictionaries for metadata relevant to the given object and return it. The returned object (some Markdown content, for example) will by default display itself intelligently. This design also makes it easy to use the doc system in a programmatic way; for example, to re-use documentation between different versions of a function:

Base.Markdown.Code

@doc "..." foo!
@doc (@doc foo!) foo

Base.Markdown.Paragraph

Or for use with Julia's metaprogramming functionality:

Base.Markdown.Code

for (f, op) in ((:add, :+), (:subtract, :-), (:multiply, :*), (:divide, :/))
    @eval begin
        $f(a,b) = $op(a,b)
    end
end
@doc "`add(a,b)` adds `a` and `b` together" add
@doc "`subtract(a,b)` subtracts `b` from `a`" subtract

Base.Markdown.Paragraph

Documentation written in non-toplevel blocks, such as if, for, and let, are not automatically added to the documentation system. @doc must be used in these cases. For example:

Base.Markdown.Code

if VERSION > v"0.4"
    "..."
    f(x) = x
end

Base.Markdown.Paragraph

will not add any documentation to f even when the condition is true and must instead be written as:

Base.Markdown.Code

if VERSION > v"0.4"
    @doc "..." ->
    f(x) = x
end

Base.Markdown.Header{2}

Syntax Guide

Base.Markdown.Paragraph

A comprehensive overview of all documentable Julia syntax.

Base.Markdown.Paragraph

In the following examples "..." is used to illustrate an arbitrary docstring which may be one of the follow four variants and contain arbitrary text:

Base.Markdown.Code

"..."

doc"..."

"""
...
"""

doc"""
...
"""

Base.Markdown.Paragraph

@doc_str should only be used when the docstring contains $ or \ characters that should not be parsed by Julia such as LaTeX syntax or Julia source code examples containing interpolation.

Base.Markdown.Header{3}

Functions and Methods

Base.Markdown.Code

"..."
function f end

"..."
f

Base.Markdown.Paragraph

Adds docstring "..." to Functionf. The first version is the preferred syntax, however both are equivalent.

Base.Markdown.Code

"..."
f(x) = x

"..."
function f(x)
    x
end

"..."
f(x)

Base.Markdown.Paragraph

Adds docstring "..." to Methodf(::Any).

Base.Markdown.Code

"..."
f(x, y = 1) = x + y

Base.Markdown.Paragraph

Adds docstring "..." to two Methods, namely f(::Any) and f(::Any, ::Any).

Base.Markdown.Header{3}

Macros

Base.Markdown.Code

"..."
macro m(x) end

Base.Markdown.Paragraph

Adds docstring "..." to the @m(::Any) macro definition.

Base.Markdown.Code

"..."
:(@m)

Base.Markdown.Paragraph

Adds docstring "..." to the macro named @m.

Base.Markdown.Header{3}

Types

Base.Markdown.Code

"..."
abstract T1

"..."
type T2
    ...
end

"..."
immutable T3
    ...
end

Base.Markdown.Paragraph

Adds the docstring "..." to types T1, T2, and T3.

Base.Markdown.Code

"..."
type T
    "x"
    x
    "y"
    y
end

Base.Markdown.Paragraph

Adds docstring "..." to type T, "x" to field T.x and "y" to field T.y. Also applicable to immutable types.

Base.Markdown.Code

"..."
typealias A T

Base.Markdown.Paragraph

Adds docstring "..." to the BindingA.

Base.Markdown.Paragraph

Bindings are used to store a reference to a particular Symbol in a Module without storing the referenced value itself.

Base.Markdown.Header{3}

Modules

Base.Markdown.Code

"..."
module M end

module M

"..."
M

end

Base.Markdown.Paragraph

Adds docstring "..." to the ModuleM. Adding the docstring above the Module is the preferred syntax, however both are equivalent.

Base.Markdown.Code

"..."
baremodule M
# ...
end

baremodule M

import Base: @doc

"..."
f(x) = x

end

Base.Markdown.Paragraph

Documenting a baremodule by placing a docstring above the expression automatically imports @doc into the module. These imports must be done manually when the module expression is not documented. Empty baremodules cannot be documented.

Base.Markdown.Header{3}

Global Variables

Base.Markdown.Code

"..."
const a = 1

"..."
b = 2

"..."
global c = 3

Base.Markdown.Paragraph

Adds docstring "..." to the Bindings a, b, and c.

Base.Markdown.Code

"..."
sym

Base.Markdown.Paragraph

Adds docstring "..." to the value associated with sym. Users should prefer documenting sym at it's definition.

Base.Markdown.Header{3}

Multiple Objects

Base.Markdown.Code

"..."
a, b

Base.Markdown.Paragraph

Adds docstring "..." to a and b each of which should be a documentable expression. This syntax is equivalent to

Base.Markdown.Code

"..."
a

"..."
b

Base.Markdown.Paragraph

Any number of expressions many be documented together in this way. This syntax can be useful when two functions are related, such as non-mutating and mutating versions f and f!.

Base.Markdown.Header{3}

Macro-generated code

Base.Markdown.Code

"..."
@m expression

Base.Markdown.Paragraph

Adds docstring "..." to expression generated by expanding @m expression. This allows for expressions decorated with @inline, @noinline, @generated, or any other macro to be documented in the same way as undecorated expressions.

Base.Markdown.Paragraph

Macro authors should take note that only macros that generate a single expression will automatically support docstrings. If a macro returns a block containing multiple subexpressions then the subexpression that should be documented must be marked using the @__doc__() macro.

Base.Markdown.Paragraph

The @enum macro makes use of @__doc__ to allow for documenting Enums. Examining it's definition should serve as an example of how to use @__doc__ correctly.

# Core.@__doc__ — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

Base.Markdown.Header{2}

Markdown Syntax Notes

Base.Markdown.Paragraph

Julia's Markdown parser supports most of the basic Markdown elements, including paragraphs, code blocks, bulleted lists and basic links. It's also a work in progress, however, and support for more advanced things like tables is in the works.

Base.Markdown.Paragraph

Markdown.jl supports interpolation in a very similar way to basic string literals, with the difference that it will store the object itself in the Markdown tree (as opposed to converting it to a string). When the Markdown content is rendered the usual show methods will be called, and these can be overridden as usual. This design allows the Markdown to be extended with arbitrarily complex features (such as references) without cluttering the basic syntax.

Base.Markdown.Paragraph

In principle, the Markdown parser itself can also be arbitrarily extended by packages, or an entirely custom flavour of Markdown can be used, but this should generally be unnecessary.