Base.Markdown.Header{1}

Arrays

Base.Markdown.Header{2}

Basic functions

Base.Markdown.Code

Base.ndims
Base.size
Base.indices
Base.indices
Base.length
Base.eachindex
Base.linearindices
Base.linearindexing
Base.countnz
Base.conj!
Base.stride
Base.strides
Base.ind2sub
Base.ind2sub
Base.sub2ind
Base.LinAlg.checksquare

Base.Markdown.Header{2}

Constructors

Base.Markdown.Code

Core.Array
Base.getindex
Base.zeros
Base.zeros
Base.ones
Base.ones
Base.trues
Base.trues
Base.falses
Base.falses
Base.fill
Base.fill!
Base.reshape
Base.similar
Base.similar
Base.reinterpret
Base.eye
Base.eye
Base.eye
Base.linspace
Base.logspace

Base.Markdown.Header{2}

Mathematical operators and functions

Base.Markdown.Paragraph

All mathematical operations and functions are supported for arrays

# Base.Broadcast.broadcast — Function.
MethodsThis function has no methods to display.Hiding 7 methods defined outside of this package.

# Base.Broadcast.broadcast! — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Broadcast.bitbroadcast — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

Base.Markdown.Header{2}

Indexing, Assignment, and Concatenation

Base.Markdown.Code

Base.getindex
Base.view
Base.@view
Base.parent
Base.parentindexes
Base.slicedim
Base.setindex!
Base.Broadcast.broadcast_getindex
Base.Broadcast.broadcast_setindex!
Base.isassigned
Base.cat
Base.vcat
Base.hcat
Base.hvcat
Base.flipdim
Base.circshift
Base.find
Base.find
Base.findn
Base.findnz
Base.findfirst
Base.findfirst
Base.findfirst
Base.findlast
Base.findlast
Base.findlast
Base.findnext
Base.findnext
Base.findnext
Base.findprev
Base.findprev
Base.findprev
Base.permutedims
Base.ipermutedims
Base.permutedims!
Base.squeeze
Base.vec
Base.promote_shape
Base.checkbounds
Base.checkbounds
Base.checkindex
Base.Random.randsubseq
Base.Random.randsubseq!

Base.Markdown.Header{2}

Array functions

Base.Markdown.Code

Base.cumprod
Base.cumprod!
Base.cumsum
Base.cumsum!
Base.cumsum_kbn
Base.cummin
Base.cummax
Base.LinAlg.diff
Base.LinAlg.gradient
Base.rot180
Base.rot180
Base.rotl90
Base.rotl90
Base.rotr90
Base.rotr90
Base.reducedim
Base.mapreducedim
Base.mapslices
Base.sum_kbn

Base.Markdown.Header{2}

Combinatorics

Base.Markdown.Code

Base.Random.randperm
Base.invperm
Base.isperm
Base.permute!
Base.ipermute!
Base.Random.randcycle
Base.Random.shuffle
Base.Random.shuffle!
Base.reverse
Base.reverseind
Base.reverse!

Base.Markdown.Header{2}

BitArrays

Base.Markdown.Paragraph

BitArrays are space-efficient "packed" boolean arrays, which store one bit per boolean value. They can be used similarly to Array{Bool} arrays (which store one byte per boolean value), and can be converted to/from the latter via Array(bitarray) and BitArray(array), respectively.

Base.Markdown.Code

Base.flipbits!
Base.rol!
Base.rol!
Base.rol
Base.ror!
Base.ror!
Base.ror

Base.Markdown.Header{2}

Sparse Vectors and Matrices

Base.Markdown.Paragraph

Sparse vectors and matrices largely support the same set of operations as their dense counterparts. The following functions are specific to sparse arrays.

Base.Markdown.Code

Base.SparseArrays.sparse
Base.SparseArrays.sparsevec
Base.SparseArrays.sparsevec
Base.SparseArrays.issparse
Base.SparseArrays.sparse
Base.SparseArrays.sparsevec
Base.full
Base.SparseArrays.nnz
Base.SparseArrays.spzeros
Base.SparseArrays.spones
Base.SparseArrays.speye
Base.SparseArrays.speye
Base.SparseArrays.spdiagm
Base.SparseArrays.sprand
Base.SparseArrays.sprandn
Base.SparseArrays.nonzeros
Base.SparseArrays.rowvals
Base.SparseArrays.nzrange
Base.SparseArrays.dropzeros!
Base.SparseArrays.dropzeros
Base.SparseArrays.dropzeros!
Base.SparseArrays.dropzeros
Base.SparseArrays.permute
Base.permute!