Essentials

Introduction

The Julia standard library contains a range of functions and macros appropriate for performing scientific and numerical computing, but is also as broad as those of many general purpose programming languages. Additional functionality is available from a growing collection of available packages. Functions are grouped by topic below.

Some general notes:

Except for functions in built-in modules (Pkg, Collections, Test and Profile), all functions documented here are directly available for use in programs.

To use module functions, use import Module to import the module, and Module.fn(x) to use the functions.

Alternatively, using Module will import all exported Module functions into the current namespace.

By convention, function names ending with an exclamation point (!) modify their arguments. Some functions have both modifying (e.g., sort!) and non-modifying (sort) versions.

Getting Around

Base.exit
Base.quit
Base.atexit
Base.atreplinit
Base.isinteractive
Base.whos
Base.summarysize
Base.edit
Base.edit
Base.@edit
Base.less
Base.less
Base.@less
Base.clipboard
Base.clipboard
Base.reload
Base.require
Base.compilecache
Base.__precompile__
Base.include_from_node1
Base.include_string
Base.include_dependency
Base.Docs.apropos
Base.which
Base.which
Base.@which
Base.methods
Base.methodswith
Base.@show
Base.versioninfo
Base.workspace
ans

All Objects

# Core.is — Function.

is(x, y) -> Bool
===(x,y) -> Bool
≡(x,y) -> Bool

Determine whether x and y are identical, in the sense that no program could distinguish them. Compares mutable objects by address in memory, and compares immutable objects (such as numbers) by contents at the bit level. This function is sometimes called egal.

source

# Core.isa — Function.

isa(x, type) -> Bool

Determine whether x is of the given type.

source

# Base.isequal — Function.

isequal(x, y)

Similar to ==, except treats all floating-point NaN values as equal to each other, and treats -0.0 as unequal to 0.0. The default implementation of isequal calls ==, so if you have a type that doesn't have these floating-point subtleties then you probably only need to define ==.

isequal is the comparison function used by hash tables (Dict). isequal(x,y) must imply that hash(x) == hash(y).

This typically means that if you define your own == function then you must define a corresponding hash (and vice versa). Collections typically implement isequal by calling isequal recursively on all contents.

Scalar types generally do not need to implement isequal separate from ==, unless they represent floating-point numbers amenable to a more efficient implementation than that provided as a generic fallback (based on isnan, signbit, and ==).

source

# Base.isless — Function.

isless(x, y)

Test whether x is less than y, according to a canonical total order. Values that are normally unordered, such as NaN, are ordered in an arbitrary but consistent fashion. This is the default comparison used by sort. Non-numeric types with a canonical total order should implement this function. Numeric types only need to implement it if they have special values such as NaN.

source

# Base.ifelse — Function.

ifelse(condition::Bool, x, y)

Return x if condition is true, otherwise return y. This differs from ? or if in that it is an ordinary function, so all the arguments are evaluated first. In some cases, using ifelse instead of an if statement can eliminate the branch in generated code and provide higher performance in tight loops.

source

# Base.lexcmp — Function.

lexcmp(x, y)

Compare x and y lexicographically and return -1, 0, or 1 depending on whether x is less than, equal to, or greater than y, respectively. This function should be defined for lexicographically comparable types, and lexless will call lexcmp by default.

source

# Base.lexless — Function.

lexless(x, y)

Determine whether x is lexicographically less than y.

source

# Core.typeof — Function.

typeof(x)

Get the concrete type of x.

source

# Core.tuple — Function.

tuple(xs...)

Construct a tuple of the given objects.

source

# Base.ntuple — Function.

ntuple(f::Function, n)

Create a tuple of length n, computing each element as f(i), where i is the index of the element.

source

# Base.object_id — Function.

object_id(x)

Get a hash value for x based on object identity. object_id(x)==object_id(y) if x === y.

source

# Base.hash — Function.

hash(x[, h::UInt])

Compute an integer hash code such that isequal(x,y) implies hash(x)==hash(y). The optional second argument h is a hash code to be mixed with the result.

New types should implement the 2-argument form, typically by calling the 2-argument hash method recursively in order to mix hashes of the contents with each other (and with h). Typically, any type that implements hash should also implement its own == (hence isequal) to guarantee the property mentioned above.

source

# Base.finalizer — Function.

finalizer(x, function)

Register a function f(x) to be called when there are no program-accessible references to x. The behavior of this function is unpredictable if x is of a bits type.

source

# Base.finalize — Function.

finalize(x)

Immediately run finalizers registered for object x.

source

# Base.copy — Function.

copy(x)

Create a shallow copy of x: the outer structure is copied, but not all internal values. For example, copying an array produces a new array with identically-same elements as the original.

source

# Base.deepcopy — Function.

deepcopy(x)

Create a deep copy of x: everything is copied recursively, resulting in a fully independent object. For example, deep-copying an array produces a new array whose elements are deep copies of the original elements. Calling deepcopy on an object should generally have the same effect as serializing and then deserializing it.

As a special case, functions can only be actually deep-copied if they are anonymous, otherwise they are just copied. The difference is only relevant in the case of closures, i.e. functions which may contain hidden internal references.

While it isn't normally necessary, user-defined types can override the default deepcopy behavior by defining a specialized version of the function deepcopy_internal(x::T, dict::ObjectIdDict) (which shouldn't otherwise be used), where T is the type to be specialized for, and dict keeps track of objects copied so far within the recursion. Within the definition, deepcopy_internal should be used in place of deepcopy, and the dict variable should be updated as appropriate before returning.

source

# Core.isdefined — Function.

isdefined([m::Module,] s::Symbol)
isdefined(object, s::Symbol)
isdefined(a::AbstractArray, index::Int)

Tests whether an assignable location is defined. The arguments can be a module and a symbol, a composite object and field name (as a symbol), or an array and index. With a single symbol argument, tests whether a global variable with that name is defined in current_module().

source

# Base.convert — Function.

convert(T, x)

Convert x to a value of type T.

If T is an Integer type, an InexactError will be raised if x is not representable by T, for example if x is not integer-valued, or is outside the range supported by T.

julia> convert(Int, 3.0)
3

julia> convert(Int, 3.5)
ERROR: InexactError()
 in convert(::Type{Int64}, ::Float64) at ./int.jl:239
 ...

If T is a AbstractFloat or Rational type, then it will return the closest value to x representable by T.

julia> x = 1/3
0.3333333333333333

julia> convert(Float32, x)
0.33333334f0

julia> convert(Rational{Int32}, x)
1//3

julia> convert(Rational{Int64}, x)
6004799503160661//18014398509481984

If T is a collection type and x a collection, the result of convert(T, x) may alias x.

julia> x = Int[1,2,3];

julia> y = convert(Vector{Int}, x);

julia> y === x
true

Similarly, if T is a composite type and x a related instance, the result of convert(T, x) may alias part or all of x.

julia> x = speye(5);

julia> typeof(x)
SparseMatrixCSC{Float64,Int64}

julia> y = convert(SparseMatrixCSC{Float64,Int64}, x);

julia> z = convert(SparseMatrixCSC{Float32,Int64}, y);

julia> y === x
true

julia> z === x
false

julia> z.colptr === x.colptr
true

source

# Base.promote — Function.

promote(xs...)

Convert all arguments to their common promotion type (if any), and return them all (as a tuple).

source

# Base.oftype — Function.

oftype(x, y)

Convert y to the type of x (convert(typeof(x), y)).

source

# Base.widen — Function.

widen(x)

If x is a type, return a "larger" type (for numeric types, this will be a type with at least as much range and precision as the argument, and usually more). Otherwise x is converted to widen(typeof(x)).

julia> widen(Int32)
Int64

julia> widen(1.5f0)
1.5

source

# Base.identity — Function.

identity(x)

The identity function. Returns its argument.

source

Types

Base.supertype
Core.issubtype
Base.:<:
Base.subtypes
Base.typemin
Base.typemax
Base.realmin
Base.realmax
Base.maxintfloat
Base.sizeof
Base.eps
Base.eps
Base.eps
Base.promote_type
Base.promote_rule
Core.getfield
Core.setfield!
Base.fieldoffset
Core.fieldtype
Base.isimmutable
Base.isbits
Base.isleaftype
Base.typejoin
Base.typeintersect
Base.Val
Base.Enums.@enum
Base.instances

Generic Functions

# Base.method_exists — Function.

method_exists(f, Tuple type) -> Bool

Determine whether the given generic function has a method matching the given Tuple of argument types.

julia> method_exists(length, Tuple{Array})
true

source

# Core.applicable — Function.

applicable(f, args...) -> Bool

Determine whether the given generic function has a method applicable to the given arguments.

julia> function f(x, y)
           x + y
       end;

julia> applicable(f, 1)
false

julia> applicable(f, 1, 2)
true

source

# Core.invoke — Function.

invoke(f, (types...), args...)

Invoke a method for the given generic function matching the specified types (as a tuple), on the specified arguments. The arguments must be compatible with the specified types. This allows invoking a method other than the most specific matching method, which is useful when the behavior of a more general definition is explicitly needed (often as part of the implementation of a more specific method of the same function).

source

# Base.:|> — Function.

|>(x, f)

Applies a function to the preceding argument. This allows for easy function chaining.

julia> [1:5;] |> x->x.^2 |> sum |> inv
0.01818181818181818

source

Syntax

Base.eval
Base.@eval
Base.evalfile
Base.esc
Base.gensym
Base.@gensym
Base.parse
Base.parse

Nullables

Base.Nullable
Base.get
Base.isnull

System

Base.run
Base.spawn
Base.DevNull
Base.success
Base.process_running
Base.process_exited
Base.kill
Base.Sys.set_process_title
Base.Sys.get_process_title
Base.readandwrite
Base.ignorestatus
Base.detach
Base.Cmd
Base.setenv
Base.withenv
Base.pipeline
Base.pipeline
Base.Libc.gethostname
Base.getipaddr
Base.Libc.getpid
Base.Libc.time
Base.time_ns
Base.tic
Base.toc
Base.toq
Base.@time
Base.@timev
Base.@timed
Base.@elapsed
Base.@allocated
Base.EnvHash
Base.ENV
Base.is_unix
Base.is_apple
Base.is_linux
Base.is_bsd
Base.is_windows
Base.Sys.windows_version
Base.@static

Errors

# Base.error — Function.

error(message::AbstractString)

Raise an ErrorException with the given message.

source

# Core.throw — Function.

throw(e)

Throw an object as an exception.

source

# Base.rethrow — Function.

rethrow([e])

Throw an object without changing the current exception backtrace. The default argument is the current exception (if called within a catch block).

source

# Base.backtrace — Function.

backtrace()

Get a backtrace object for the current program point.

source

# Base.catch_backtrace — Function.

catch_backtrace()

Get the backtrace of the current exception, for use within catch blocks.

source

# Base.assert — Function.

assert(cond)

Throw an AssertionError if cond is false. Also available as the macro @assert expr.

source

# Base.@assert — Macro.

@assert cond [text]

Throw an AssertionError if cond is false. Preferred syntax for writing assertions. Message text is optionally displayed upon assertion failure.

source

# Base.ArgumentError — Type.

ArgumentError(msg)

The parameters to a function call do not match a valid signature. Argument msg is a descriptive error string.

source

# Base.AssertionError — Type.

AssertionError([msg])

The asserted condition did not evaluate to true. Optional argument msg is a descriptive error string.

source

# Core.BoundsError — Type.

BoundsError([a],[i])

An indexing operation into an array, a, tried to access an out-of-bounds element, i.

source

# Base.DimensionMismatch — Type.

DimensionMismatch([msg])

The objects called do not have matching dimensionality. Optional argument msg is a descriptive error string.

source

# Core.DivideError — Type.

DivideError()

Integer division was attempted with a denominator value of 0.

source

# Core.DomainError — Type.

DomainError()

The arguments to a function or constructor are outside the valid domain.

source

# Base.EOFError — Type.

EOFError()

No more data was available to read from a file or stream.

source

# Core.ErrorException — Type.

ErrorException(msg)

Generic error type. The error message, in the .msg field, may provide more specific details.

source

# Core.InexactError — Type.

InexactError()

Type conversion cannot be done exactly.

source

# Core.InterruptException — Type.

InterruptException()

The process was stopped by a terminal interrupt (CTRL+C).

source

# Base.KeyError — Type.

KeyError(key)

An indexing operation into an Associative (Dict) or Set like object tried to access or delete a non-existent element.

source

# Base.LoadError — Type.

LoadError(file::AbstractString, line::Int, error)

An error occurred while includeing, requireing, or using a file. The error specifics should be available in the .error field.

source

# Base.MethodError — Type.

MethodError(f, args)

A method with the required type signature does not exist in the given generic function. Alternatively, there is no unique most-specific method.

source

# Base.NullException — Type.

NullException()

An attempted access to a Nullable with no defined value.

source

# Core.OutOfMemoryError — Type.

OutOfMemoryError()

An operation allocated too much memory for either the system or the garbage collector to handle properly.

source

# Core.ReadOnlyMemoryError — Type.

ReadOnlyMemoryError()

An operation tried to write to memory that is read-only.

source

# Core.OverflowError — Type.

OverflowError()

The result of an expression is too large for the specified type and will cause a wraparound.

source

# Base.ParseError — Type.

ParseError(msg)

The expression passed to the parse function could not be interpreted as a valid Julia expression.

source

# Base.ProcessExitedException — Type.

ProcessExitedException()

After a client Julia process has exited, further attempts to reference the dead child will throw this exception.

source

# Core.StackOverflowError — Type.

StackOverflowError()

The function call grew beyond the size of the call stack. This usually happens when a call recurses infinitely.

source

# Base.SystemError — Type.

SystemError(prefix::AbstractString, [errno::Int32])

A system call failed with an error code (in the errno global variable).

source

# Core.TypeError — Type.

TypeError(func::Symbol, context::AbstractString, expected::Type, got)

A type assertion failure, or calling an intrinsic function with an incorrect argument type.

source

# Core.UndefRefError — Type.

UndefRefError()

The item or field is not defined for the given object.

source

# Core.UndefVarError — Type.

UndefVarError(var::Symbol)

A symbol in the current scope is not defined.

source

# Base.InitError — Type.

InitError(mod::Symbol, error)

An error occurred when running a module's __init__ function. The actual error thrown is available in the .error field.

source

# Base.retry — Function.

retry(f, [retry_on]; n=1, max_delay=10.0) -> Function

Returns a lambda that retries function f up to n times in the event of an exception. If retry_on is a Type then retry only for exceptions of that type. If retry_on is a function test_error(::Exception) -> Bool then retry only if it is true.

The first retry happens after a gap of 50 milliseconds or max_delay, whichever is lower. Subsequently, the delays between retries are exponentially increased with a random factor up to max_delay.

Examples

retry(http_get, e -> e.status == "503")(url)
retry(read, UVError)(io)

source

Events

Base.Timer
Base.Timer
Base.AsyncCondition
Base.AsyncCondition

Reflection

Base.module_name
Base.module_parent
Base.current_module
Base.fullname
Base.names
Core.nfields
Base.fieldnames
Base.fieldname
Base.datatype_module
Base.isconst
Base.function_name
Base.function_module
Base.function_module
Base.functionloc
Base.functionloc
Base.@functionloc

Internals

# Base.gc — Function.

gc()

Perform garbage collection. This should not generally be used.

source

# Base.gc_enable — Function.

gc_enable(on::Bool)

Control whether garbage collection is enabled using a boolean argument (true for enabled, false for disabled). Returns previous GC state. Disabling garbage collection should be used only with extreme caution, as it can cause memory use to grow without bound.

source

# Base.macroexpand — Function.

macroexpand(x)

Takes the expression x and returns an equivalent expression with all macros removed (expanded).

source

# Base.expand — Function.

expand(x)

Takes the expression x and returns an equivalent expression in lowered form.

source

# Base.code_lowered — Function.

code_lowered(f, types)

Returns an array of lowered ASTs for the methods matching the given generic function and type signature.

source

# Base.@code_lowered — Macro.

@code_lowered

Evaluates the arguments to the function or macro call, determines their types, and calls code_lowered on the resulting expression.

source

# Base.code_typed — Function.

code_typed(f, types; optimize=true)

Returns an array of lowered and type-inferred ASTs for the methods matching the given generic function and type signature. The keyword argument optimize controls whether additional optimizations, such as inlining, are also applied.

source

# Base.@code_typed — Macro.

@code_typed

Evaluates the arguments to the function or macro call, determines their types, and calls code_typed on the resulting expression.

source

# Base.code_warntype — Function.

code_warntype([io], f, types)

Prints lowered and type-inferred ASTs for the methods matching the given generic function and type signature to io which defaults to STDOUT. The ASTs are annotated in such a way as to cause "non-leaf" types to be emphasized (if color is available, displayed in red). This serves as a warning of potential type instability. Not all non-leaf types are particularly problematic for performance, so the results need to be used judiciously. See Manual for more information.

source

# Base.@code_warntype — Macro.

@code_warntype

Evaluates the arguments to the function or macro call, determines their types, and calls code_warntype on the resulting expression.

source

# Base.code_llvm — Function.

code_llvm([io], f, types)

Prints the LLVM bitcodes generated for running the method matching the given generic function and type signature to io which defaults to STDOUT.

All metadata and dbg.* calls are removed from the printed bitcode. Use code_llvm_raw for the full IR.

source

# Base.@code_llvm — Macro.

@code_llvm

Evaluates the arguments to the function or macro call, determines their types, and calls code_llvm on the resulting expression.

source

# Base.code_native — Function.

code_native([io], f, types)

Prints the native assembly instructions generated for running the method matching the given generic function and type signature to io which defaults to STDOUT.

source

# Base.@code_native — Macro.

@code_native

Evaluates the arguments to the function or macro call, determines their types, and calls code_native on the resulting expression.

source

# Base.precompile — Function.

precompile(f,args::Tuple{Vararg{Any}})

Compile the given function f for the argument tuple (of types) args, but do not execute it.

source