Base.Markdown.Header{1}

Collections and Data Structures

Base.Markdown.Header{2}

Iteration

Base.Markdown.Paragraph

Sequential iteration is implemented by the methods start(), done(), and next(). The general for loop:

Base.Markdown.Code

for i = I   # or  "for i in I"
    # body
end

Base.Markdown.Paragraph

is translated into:

Base.Markdown.Code

state = start(I)
while !done(I, state)
    (i, state) = next(I, state)
    # body
end

Base.Markdown.Paragraph

The state object may be anything, and should be chosen appropriately for each iterable type. See the manual section on the iteration interface for more details about defining a custom iterable type.

# Base.start — Function.
MethodsThis function has no methods to display.Hiding 59 methods defined outside of this package.

# Base.done — Function.
MethodsThis function has no methods to display.Hiding 56 methods defined outside of this package.

# Base.next — Function.
MethodsThis function has no methods to display.Hiding 71 methods defined outside of this package.

# Base.zip — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.enumerate — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.rest — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.countfrom — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.take — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.drop — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.cycle — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.repeated — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.iteratorsize — Function.
MethodsThis function has no methods to display.Hiding 28 methods defined outside of this package.

# Base.iteratoreltype — Function.
MethodsThis function has no methods to display.Hiding 18 methods defined outside of this package.

Base.Markdown.Paragraph

Fully implemented by:

Base.Markdown.List

Range

UnitRange

Tuple

Number

AbstractArray

IntSet

ObjectIdDict

Dict

WeakKeyDict

EachLine

AbstractString

Set

Task

Base.Markdown.Header{2}

General Collections

# Base.isempty — Function.
MethodsThis function has no methods to display.Hiding 29 methods defined outside of this package.

# Base.empty! — Function.
MethodsThis function has no methods to display.Hiding 7 methods defined outside of this package.

# Base.length — Function.
MethodsThis function has no methods to display.Hiding 64 methods defined outside of this package.

# Base.endof — Function.
MethodsThis function has no methods to display.Hiding 14 methods defined outside of this package.

Base.Markdown.Paragraph

Fully implemented by:

Base.Markdown.List

Range

UnitRange

Tuple

Number

AbstractArray

IntSet

Dict

WeakKeyDict

AbstractString

Set

Base.Markdown.Header{2}

Iterable Collections

Base.Markdown.Code

Base.in
Base.eltype
Base.indexin
Base.findin
Base.unique
Base.unique
Base.unique
Base.allunique
Base.reduce
Base.reduce
Base.foldl
Base.foldl
Base.foldr
Base.foldr
Base.maximum
Base.maximum
Base.maximum!
Base.minimum
Base.minimum
Base.minimum!
Base.extrema
Base.extrema
Base.indmax
Base.indmin
Base.findmax
Base.findmax
Base.findmin
Base.findmin
Base.findmax!
Base.findmin!
Base.maxabs
Base.maxabs
Base.maxabs!
Base.minabs
Base.minabs
Base.minabs!
Base.sum
Base.sum
Base.sum!
Base.sum
Base.sumabs
Base.sumabs
Base.sumabs!
Base.sumabs2
Base.sumabs2
Base.sumabs2!
Base.prod
Base.prod
Base.prod!
Base.any
Base.any
Base.any!
Base.all
Base.all
Base.all!
Base.count
Base.any
Base.all
Base.foreach
Base.map
Base.map!
Base.map!
Base.mapreduce
Base.mapreduce
Base.mapfoldl
Base.mapfoldl
Base.mapfoldr
Base.mapfoldr
Base.first
Base.last
Base.step
Base.collect
Base.collect
Base.issubset
Base.filter
Base.filter!

Base.Markdown.Header{2}

Indexable Collections

# Base.getindex — Function.
Methods

getindex(
    t::Documenter.Utilities.DOM.Tag,
    args...
)

defined at src/Utilities/DOM.jl:198

getindex(
    n::Documenter.Utilities.DOM.Node,
    args...
)

defined at src/Utilities/DOM.jl:199

Hiding 170 methods defined outside of this package.

# Base.setindex! — Function.
MethodsThis function has no methods to display.Hiding 95 methods defined outside of this package.

Base.Markdown.Paragraph

Fully implemented by:

Base.Markdown.List

Array

BitArray

AbstractArray

SubArray

ObjectIdDict

Dict

WeakKeyDict

AbstractString

Base.Markdown.Paragraph

Partially implemented by:

Base.Markdown.List

Range

UnitRange

Tuple

Base.Markdown.Header{2}

Associative Collections

Base.Markdown.Paragraph

Dict is the standard associative collection. Its implementation uses hash() as the hashing function for the key, and isequal() to determine equality. Define these two functions for custom types to override how they are stored in a hash table.

Base.Markdown.Paragraph

ObjectIdDict is a special hash table where the keys are always object identities.

Base.Markdown.Paragraph

WeakKeyDict is a hash table implementation where the keys are weak references to objects, and thus may be garbage collected even when referenced in a hash table.

Base.Markdown.Paragraph

Dicts can be created by passing pair objects constructed with =>() to a Dict constructor: Dict("A"=>1, "B"=>2). This call will attempt to infer type information from the keys and values (i.e. this example creates a Dict{String, Int64}). To explicitly specify types use the syntax Dict{KeyType,ValueType}(...). For example, Dict{String,Int32}("A"=>1, "B"=>2).

Base.Markdown.Paragraph

Dicts may also be created with generators. For example, Dict(i => f(i) for i = 1:10).

Base.Markdown.Paragraph

Given a dictionary D, the syntax D[x] returns the value of key x (if it exists) or throws an error, and D[x] = y stores the key-value pair x => y in D (replacing any existing value for the key x). Multiple arguments to D[...] are converted to tuples; for example, the syntax D[x,y] is equivalent to D[(x,y)], i.e. it refers to the value keyed by the tuple (x,y).

Base.Markdown.Code

Base.Dict
Base.haskey
Base.get
Base.get
Base.get!
Base.get!
Base.getkey
Base.delete!
Base.pop!
Base.keys
Base.values
Base.merge
Base.merge!
Base.sizehint!
Base.keytype
Base.valtype

Base.Markdown.Paragraph

Fully implemented by:

Base.Markdown.List

ObjectIdDict

Dict

WeakKeyDict

Base.Markdown.Paragraph

Partially implemented by:

Base.Markdown.List

IntSet

Set

EnvHash

Array

BitArray

Base.Markdown.Header{2}

Set-Like Collections

Base.Markdown.Code

Base.Set
Base.IntSet
Base.union
Base.union!
Base.intersect
Base.setdiff
Base.setdiff!
Base.symdiff
Base.symdiff!
Base.symdiff!
Base.symdiff!
Base.intersect!
Base.issubset

Base.Markdown.Paragraph

Fully implemented by:

Base.Markdown.List

IntSet

Set

Base.Markdown.Paragraph

Partially implemented by:

Base.Markdown.List

Array

Base.Markdown.Header{2}

Dequeues

Base.Markdown.Code

Base.push!
Base.pop!
Base.unshift!
Base.shift!
Base.insert!
Base.deleteat!
Base.deleteat!
Base.splice!
Base.splice!
Base.resize!
Base.append!
Base.prepend!

Base.Markdown.Paragraph

Fully implemented by:

Base.Markdown.List

Vector (a.k.a. 1-dimensional Array)

BitVector (a.k.a. 1-dimensional BitArray)

Base.Markdown.Header{2}

PriorityQueue

Base.Markdown.Paragraph

The PriorityQueue type is available from the Collections module. It provides a basic priority queue implementation allowing for arbitrary key and priority types. Multiple identical keys are not permitted, but the priority of existing keys can be changed efficiently.

# Base.Collections.PriorityQueue — Type.
MethodsThis function has no methods to display.Hiding 11 methods defined outside of this package.

# Base.Collections.enqueue! — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Collections.dequeue! — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Collections.peek — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

Base.Markdown.Paragraph

PriorityQueue also behaves similarly to a Dict in that keys can be inserted and priorities accessed or changed using indexing notation.

Base.Markdown.Code

julia> # Julia code

       pq = Collections.PriorityQueue();

julia> # Insert keys with associated priorities

       pq["a"] = 10; pq["b"] = 5; pq["c"] = 15; pq
Base.Collections.PriorityQueue{Any,Any,Base.Order.ForwardOrdering} with 3 entries:
  "c" => 15
  "b" => 5
  "a" => 10

julia> # Change the priority of an existing key

       pq["a"] = 0; pq
Base.Collections.PriorityQueue{Any,Any,Base.Order.ForwardOrdering} with 3 entries:
  "c" => 15
  "b" => 5
  "a" => 0

Base.Markdown.Header{2}

Heap Functions

Base.Markdown.Paragraph

Along with the PriorityQueue type, the Collections module provides lower level functions for performing binary heap operations on arrays. Each function takes an optional ordering argument. If not given, default ordering is used, so that elements popped from the heap are given in ascending order.

# Base.Collections.heapify — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Collections.heapify! — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Collections.isheap — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Collections.heappush! — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Collections.heappop! — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.