Base.Markdown.Header{1}

Dates and Time

Base.Markdown.Header{2}

Dates and Time Types

# Base.Dates.Period — Type.

# Base.Dates.CompoundPeriod — Type.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.Instant — Type.

# Base.Dates.UTInstant — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.TimeType — Type.

# Base.Dates.DateTime — Type.
MethodsThis function has no methods to display.Hiding 38 methods defined outside of this package.

# Base.Dates.Date — Type.
MethodsThis function has no methods to display.Hiding 23 methods defined outside of this package.

Base.Markdown.Header{2}

Dates Functions

Base.Markdown.Paragraph

All Dates functions are defined in the Dates module; note that only the Date, DateTime, and now functions are exported; to use all other Dates functions, you'll need to prefix each function call with an explicit Dates., e.g. Dates.dayofweek(dt). Alternatively, you can write using Base.Dates to bring all exported functions into Main to be used without the Dates. prefix.

Base.Markdown.Code

Base.Dates.DateTime
Base.Dates.DateTime
Base.Dates.DateTime
Base.Dates.DateTime
Base.Dates.DateTime
Base.Dates.format
Base.Dates.DateFormat
Base.Dates.DateTime
Base.Dates.Date
Base.Dates.Date
Base.Dates.Date
Base.Dates.Date
Base.Dates.Date
Base.Dates.Date
Base.Dates.now
Base.Dates.now
Base.eps

Base.Markdown.Header{3}

Accessor Functions

# Base.Dates.year — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.month — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.week — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.day — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.hour — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.minute — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.second — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.millisecond — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.Year — Type.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Base.Dates.Month — Type.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Base.Dates.Week — Type.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Base.Dates.Day — Type.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Base.Dates.Hour — Type.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Base.Dates.Minute — Type.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Base.Dates.Second — Type.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Base.Dates.Millisecond — Type.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Base.Dates.yearmonth — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.monthday — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.yearmonthday — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

Base.Markdown.Header{3}

Query Functions

# Base.Dates.dayname — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.dayabbr — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.dayofweek — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.dayofmonth — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.dayofweekofmonth — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.daysofweekinmonth — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.monthname — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.monthabbr — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.daysinmonth — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.isleapyear — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.dayofyear — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.daysinyear — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Dates.quarterofyear — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.dayofquarter — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

Base.Markdown.Header{3}

Adjuster Functions

Base.Markdown.Code

Base.trunc
Base.Dates.firstdayofweek
Base.Dates.lastdayofweek
Base.Dates.firstdayofmonth
Base.Dates.lastdayofmonth
Base.Dates.firstdayofyear
Base.Dates.lastdayofyear
Base.Dates.firstdayofquarter
Base.Dates.lastdayofquarter
Base.Dates.tonext
Base.Dates.toprev
Base.Dates.tofirst
Base.Dates.tolast
Base.Dates.tonext
Base.Dates.toprev
Base.Dates.recur

Base.Markdown.Header{3}

Periods

Base.Markdown.Code

Base.Dates.Period
Base.Dates.CompoundPeriod
Base.Dates.default

Base.Markdown.Header{3}

Rounding Functions

Base.Markdown.Paragraph

Date and DateTime values can be rounded to a specified resolution (e.g., 1 month or 15 minutes) with floor, ceil, or round.

# Base.floor — Function.
MethodsThis function has no methods to display.Hiding 41 methods defined outside of this package.

# Base.ceil — Function.
MethodsThis function has no methods to display.Hiding 36 methods defined outside of this package.

# Base.round — Function.
MethodsThis function has no methods to display.Hiding 67 methods defined outside of this package.

Base.Markdown.Paragraph

The following functions are not exported:

# Base.Dates.floorceil — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Dates.epochdays2date — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Dates.epochms2datetime — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Dates.date2epochdays — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Dates.datetime2epochms — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

Base.Markdown.Header{3}

Conversion Functions

# Base.Dates.today — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Dates.unix2datetime — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.datetime2unix — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.julian2datetime — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.datetime2julian — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.rata2datetime — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Dates.datetime2rata — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

Base.Markdown.Header{3}

Constants

Base.Markdown.Paragraph

Days of the Week:

Base.Markdown.Table

Variable	Abbr.	Value (Int)

Monday	Mon	1

Tuesday	Tue	2

Wednesday	Wed	3

Thursday	Thu	4

Friday	Fri	5

Saturday	Sat	6

Sunday	Sun	7

Base.Markdown.Paragraph

Months of the Year:

Base.Markdown.Table

Variable	Abbr.	Value (Int)

January	Jan	1

February	Feb	2

March	Mar	3

April	Apr	4

May	May	5

June	Jun	6

July	Jul	7

August	Aug	8

September	Sep	9

October	Oct	10

November	Nov	11

December	Dec	12