Base.Markdown.Header{1}

Mathematics

Base.Markdown.Header{2}

Mathematical Operators

Base.Markdown.Code

Base.:-
Base.:+
Base.:-
Base.:*
Base.:/
Base.:\
Base.:^
Base.:.+
Base.:.-
Base.:.*
Base.:./
Base.:.\
Base.:.^
Base.fma
Base.muladd
Base.div
Base.fld
Base.cld
Base.mod
Base.Math.mod2pi
Base.rem
Base.divrem
Base.fldmod
Base.fld1
Base.mod1
Base.fldmod1
Base.://
Base.rationalize
Base.num
Base.den
Base.:<<
Base.:>>
Base.:>>>
Base.colon
Base.colon
Base.range
Base.:==
Base.:!=
Core.:===
Base.:!==
Base.:<
Base.:<=
Base.:>
Base.:>=
Base.:.==
Base.:.!=
Base.:.<
Base.:.<=
Base.:.>
Base.:.>=
Base.cmp
Base.:~
Base.:&
Base.:|
Base.:$
Base.:!
&&
||

Base.Markdown.Header{2}

Mathematical Functions

Base.Markdown.Code

Base.isapprox
Base.sin
Base.cos
Base.tan
Base.Math.sind
Base.Math.cosd
Base.Math.tand
Base.Math.sinpi
Base.Math.cospi
Base.sinh
Base.cosh
Base.tanh
Base.asin
Base.acos
Base.atan
Base.Math.atan2
Base.Math.asind
Base.Math.acosd
Base.Math.atand
Base.Math.sec
Base.Math.csc
Base.Math.cot
Base.Math.secd
Base.Math.cscd
Base.Math.cotd
Base.Math.asec
Base.Math.acsc
Base.Math.acot
Base.Math.asecd
Base.Math.acscd
Base.Math.acotd
Base.Math.sech
Base.Math.csch
Base.Math.coth
Base.asinh
Base.acosh
Base.atanh
Base.Math.asech
Base.Math.acsch
Base.Math.acoth
Base.Math.sinc
Base.Math.cosc
Base.Math.deg2rad
Base.Math.rad2deg
Base.Math.hypot
Base.Math.hypot
Base.log
Base.log
Base.log2
Base.log10
Base.log1p
Base.Math.frexp
Base.exp
Base.exp2
Base.exp10
Base.Math.ldexp
Base.Math.modf
Base.expm1
Base.round
Base.Rounding.RoundingMode
Base.Rounding.RoundNearest
Base.Rounding.RoundNearestTiesAway
Base.Rounding.RoundNearestTiesUp
Base.Rounding.RoundToZero
Base.Rounding.RoundUp
Base.Rounding.RoundDown
Base.round
Base.ceil
Base.floor
Base.trunc
Base.unsafe_trunc
Base.signif
Base.min
Base.max
Base.minmax
Base.Math.clamp
Base.Math.clamp!
Base.abs
Base.Checked.checked_abs
Base.Checked.checked_neg
Base.Checked.checked_add
Base.Checked.checked_sub
Base.Checked.checked_mul
Base.Checked.checked_div
Base.Checked.checked_rem
Base.Checked.checked_fld
Base.Checked.checked_mod
Base.Checked.checked_cld
Base.abs2
Base.copysign
Base.sign
Base.signbit
Base.flipsign
Base.sqrt
Base.isqrt
Base.Math.cbrt
Base.Math.erf
Base.Math.erfc
Base.Math.erfcx
Base.Math.erfi
Base.Math.dawson
Base.Math.erfinv
Base.Math.erfcinv
Base.real
Base.imag
Base.reim
Base.conj
Base.angle
Base.cis
Base.binomial
Base.factorial
Base.gcd
Base.lcm
Base.gcdx
Base.ispow2
Base.nextpow2
Base.prevpow2
Base.nextpow
Base.prevpow
Base.nextprod
Base.invmod
Base.powermod
Base.Math.gamma
Base.Math.lgamma
Base.Math.lfact
Base.Math.digamma
Base.Math.invdigamma
Base.Math.trigamma
Base.Math.polygamma
Base.Math.airy
Base.Math.airyai
Base.Math.airyprime
Base.Math.airyaiprime
Base.Math.airybi
Base.Math.airybiprime
Base.Math.airyx
Base.Math.besselj0
Base.Math.besselj1
Base.Math.besselj
Base.Math.besseljx
Base.Math.bessely0
Base.Math.bessely1
Base.Math.bessely
Base.Math.besselyx
Base.Math.hankelh1
Base.Math.hankelh1x
Base.Math.hankelh2
Base.Math.hankelh2x
Base.Math.besselh
Base.Math.besselhx
Base.Math.besseli
Base.Math.besselix
Base.Math.besselk
Base.Math.besselkx
Base.Math.beta
Base.Math.lbeta
Base.Math.eta
Base.Math.zeta
Base.Math.zeta
Base.ndigits
Base.widemul
Base.Math.@evalpoly

Base.Markdown.Header{2}

Statistics

Base.Markdown.Code

Base.mean
Base.mean
Base.mean!
Base.std
Base.stdm
Base.var
Base.varm
Base.middle
Base.middle
Base.middle
Base.middle
Base.median
Base.median!
Base.midpoints
Base.quantile
Base.quantile!
Base.cov
Base.cov
Base.cov
Base.cov
Base.cor
Base.cor
Base.cor
Base.cor

Base.Markdown.Header{2}

Signal Processing

Base.Markdown.Paragraph

Fast Fourier transform (FFT) functions in Julia are implemented by calling functions from FFTW. By default, Julia does not use multi-threaded FFTW. Higher performance may be obtained by experimenting with multi-threading. Use FFTW.set_num_threads(np) to use np threads.

Base.Markdown.Code

Base.DFT.fft
Base.DFT.fft!
Base.DFT.ifft
Base.DFT.ifft!
Base.DFT.bfft
Base.DFT.bfft!
Base.DFT.plan_fft
Base.DFT.plan_ifft
Base.DFT.plan_bfft
Base.DFT.plan_fft!
Base.DFT.plan_ifft!
Base.DFT.plan_bfft!
Base.DFT.rfft
Base.DFT.irfft
Base.DFT.brfft
Base.DFT.plan_rfft
Base.DFT.plan_brfft
Base.DFT.plan_irfft
Base.DFT.FFTW.dct
Base.DFT.FFTW.dct!
Base.DFT.FFTW.idct
Base.DFT.FFTW.idct!
Base.DFT.FFTW.plan_dct
Base.DFT.FFTW.plan_dct!
Base.DFT.FFTW.plan_idct
Base.DFT.FFTW.plan_idct!
Base.DFT.fftshift
Base.DFT.fftshift
Base.DFT.ifftshift
Base.DSP.filt
Base.DSP.filt!
Base.DSP.deconv
Base.DSP.conv
Base.DSP.conv2
Base.DSP.conv2
Base.DSP.xcorr

Base.Markdown.Paragraph

The following functions are defined within the Base.FFTW module.

# Base.DFT.FFTW.r2r — Function.
MethodsThis function has no methods to display.Hiding 6 methods defined outside of this package.

# Base.DFT.FFTW.r2r! — Function.
MethodsThis function has no methods to display.Hiding 6 methods defined outside of this package.

# Base.DFT.FFTW.plan_r2r — Function.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Base.DFT.FFTW.plan_r2r! — Function.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

Base.Markdown.Header{2}

Numerical Integration

Base.Markdown.Paragraph

Although several external packages are available for numeric integration and solution of ordinary differential equations, we also provide some built-in integration support in Julia.

# Base.QuadGK.quadgk — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.