Base.Markdown.Header{1}

Tasks and Parallel Computing

Base.Markdown.Header{2}

Tasks

Base.Markdown.Code

Core.Task
Base.yieldto
Base.current_task
Base.istaskdone
Base.istaskstarted
Base.consume
Base.produce
Base.yield
Base.task_local_storage
Base.task_local_storage
Base.task_local_storage
Base.Condition
Base.notify
Base.schedule
Base.@schedule
Base.@task
Base.sleep
Base.ReentrantLock
Base.lock
Base.unlock
Base.Channel

Base.Markdown.Header{2}

General Parallel Computing Support

Base.Markdown.Code

Base.addprocs
Base.addprocs
Base.addprocs
Base.addprocs
Base.nprocs
Base.nworkers
Base.procs
Base.workers
Base.rmprocs
Base.interrupt
Base.myid
Base.asyncmap
Base.pmap
Base.remotecall
Base.Future
Base.Future
Base.RemoteChannel
Base.RemoteChannel
Base.RemoteChannel
Base.wait
Base.fetch
Base.remotecall_wait
Base.remotecall_fetch
Base.put!
Base.put!
Base.put!
Base.take!
Base.take!
Base.isready
Base.isready
Base.close
Base.WorkerPool
Base.CachingPool
Base._default_worker_pool
Base.remote
Base.remotecall
Base.remotecall_wait
Base.remotecall_fetch
Base.timedwait
Base.@spawn
Base.@spawnat
Base.@fetch
Base.@fetchfrom
Base.@async
Base.@sync
Base.@parallel
Base.@everywhere
Base.clear!
Base.remoteref_id
Base.channel_from_id
Base.worker_id_from_socket
Base.cluster_cookie

Base.Markdown.Header{2}

Shared Arrays

Base.Markdown.Code

Base.SharedArray
Base.SharedArray
Base.procs
Base.sdata
Base.indexpids
Base.localindexes

Base.Markdown.Header{2}

Multi-Threading

Base.Markdown.Paragraph

This experimental interface supports Julia's multi-threading capabilities. Types and function described here might (and likely will) change in the future.

# Base.Threads.threadid — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Threads.nthreads — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Threads.@threads — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Threads.Atomic — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Threads.atomic_cas! — Function.
MethodsThis function has no methods to display.Hiding 13 methods defined outside of this package.

# Base.Threads.atomic_xchg! — Function.
MethodsThis function has no methods to display.Hiding 13 methods defined outside of this package.

# Base.Threads.atomic_add! — Function.
MethodsThis function has no methods to display.Hiding 11 methods defined outside of this package.

# Base.Threads.atomic_sub! — Function.
MethodsThis function has no methods to display.Hiding 11 methods defined outside of this package.

# Base.Threads.atomic_and! — Function.
MethodsThis function has no methods to display.Hiding 10 methods defined outside of this package.

# Base.Threads.atomic_nand! — Function.
MethodsThis function has no methods to display.Hiding 10 methods defined outside of this package.

# Base.Threads.atomic_or! — Function.
MethodsThis function has no methods to display.Hiding 10 methods defined outside of this package.

# Base.Threads.atomic_xor! — Function.
MethodsThis function has no methods to display.Hiding 10 methods defined outside of this package.

# Base.Threads.atomic_max! — Function.
MethodsThis function has no methods to display.Hiding 11 methods defined outside of this package.

# Base.Threads.atomic_min! — Function.
MethodsThis function has no methods to display.Hiding 11 methods defined outside of this package.

# Base.Threads.atomic_fence — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

Base.Markdown.Header{2}

Cluster Manager Interface

Base.Markdown.Paragraph

This interface provides a mechanism to launch and manage Julia workers on different cluster environments. LocalManager, for launching additional workers on the same host and SSHManager, for launching on remote hosts via ssh are present in Base. TCP/IP sockets are used to connect and transport messages between processes. It is possible for Cluster Managers to provide a different transport.

Base.Markdown.Code

Base.launch
Base.manage
Base.kill
Base.init_worker
Base.connect
Base.process_messages