Profiling

Base.Profile.@profile — Macro.

@profile

@profile <expression> runs your expression while taking periodic backtraces. These are appended to an internal buffer of backtraces.

source

The methods in Base.Profile are not exported and need to be called e.g. as Profile.print().

Base.Profile.clear
Base.Profile.print
Base.Profile.print
Base.Profile.init
Base.Profile.fetch
Base.Profile.retrieve
Base.Profile.callers
Base.Profile.clear_malloc_data