Arrays

Basic functions

Base.ndims
Base.size
Base.indices
Base.indices
Base.length
Base.eachindex
Base.linearindices
Base.linearindexing
Base.countnz
Base.conj!
Base.stride
Base.strides
Base.ind2sub
Base.ind2sub
Base.sub2ind
Base.LinAlg.checksquare

Constructors

Core.Array
Base.getindex
Base.zeros
Base.zeros
Base.ones
Base.ones
Base.trues
Base.trues
Base.falses
Base.falses
Base.fill
Base.fill!
Base.reshape
Base.similar
Base.similar
Base.reinterpret
Base.eye
Base.eye
Base.eye
Base.linspace
Base.logspace

Mathematical operators and functions

All mathematical operations and functions are supported for arrays

# Base.Broadcast.broadcast — Function.

broadcast(f, As...)

Broadcasts the arrays As to a common size by expanding singleton dimensions, and returns an array of the results f(as...) for each position.

source

# Base.Broadcast.broadcast! — Function.

broadcast!(f, dest, As...)

Like broadcast, but store the result of broadcast(f, As...) in the dest array. Note that dest is only used to store the result, and does not supply arguments to f unless it is also listed in the As, as in broadcast!(f, A, A, B) to perform A[:] = broadcast(f, A, B).

source

# Base.Broadcast.bitbroadcast — Function.

bitbroadcast(f, As...)

Like broadcast, but allocates a BitArray to store the result, rather then an Array.

source

Indexing, Assignment, and Concatenation

Base.getindex
Base.view
Base.@view
Base.parent
Base.parentindexes
Base.slicedim
Base.setindex!
Base.Broadcast.broadcast_getindex
Base.Broadcast.broadcast_setindex!
Base.isassigned
Base.cat
Base.vcat
Base.hcat
Base.hvcat
Base.flipdim
Base.circshift
Base.find
Base.find
Base.findn
Base.findnz
Base.findfirst
Base.findfirst
Base.findfirst
Base.findlast
Base.findlast
Base.findlast
Base.findnext
Base.findnext
Base.findnext
Base.findprev
Base.findprev
Base.findprev
Base.permutedims
Base.ipermutedims
Base.permutedims!
Base.squeeze
Base.vec
Base.promote_shape
Base.checkbounds
Base.checkbounds
Base.checkindex
Base.Random.randsubseq
Base.Random.randsubseq!

Array functions

Base.cumprod
Base.cumprod!
Base.cumsum
Base.cumsum!
Base.cumsum_kbn
Base.cummin
Base.cummax
Base.LinAlg.diff
Base.LinAlg.gradient
Base.rot180
Base.rot180
Base.rotl90
Base.rotl90
Base.rotr90
Base.rotr90
Base.reducedim
Base.mapreducedim
Base.mapslices
Base.sum_kbn

Combinatorics

# Base.Random.randperm — Function.

randperm([rng,] n)

Construct a random permutation of length n. The optional rng argument specifies a random number generator (see Random Numbers). To randomly permute a arbitrary vector, see shuffle or shuffle!.

source

# Base.invperm — Function.

invperm(v)

Return the inverse permutation of v

source

# Base.isperm — Function.

isperm(v) -> Bool

Returns true if v is a valid permutation.

source

# Base.permute! — Function.

permute!{Tv,Ti,Tp<:Integer,Tq<:Integer}(X::SparseMatrixCSC{Tv,Ti}, A::SparseMatrixCSC{Tv,Ti},
    p::AbstractVector{Tp}, q::AbstractVector{Tq}[, C::SparseMatrixCSC{Tv,Ti}])

Bilaterally permute A, storing result PAQ (A[p,q]) in X. Stores intermediate result (AQ)^T (transpose(A[:,q])) in optional argument C if present. Requires that none of X, A, and, if present, C alias each other; to store result PAQ back into A, use the following method lacking X:

permute!{Tv,Ti,Tp<:Integer,Tq<:Integer}(A::SparseMatrixCSC{Tv,Ti}, p::AbstractVector{Tp},
    q::AbstractVector{Tq}[, C::SparseMatrixCSC{Tv,Ti}[, workcolptr::Vector{Ti}]])

X's dimensions must match those of A (X.m == A.m and X.n == A.n), and X must have enough storage to accommodate all allocated entries in A (length(X.rowval) >= nnz(A) and length(X.nzval) >= nnz(A)). Column-permutation q's length must match A's column count (length(q) == A.n). Row-permutation p's length must match A's row count (length(p) == A.m).

C's dimensions must match those of transpose(A) (C.m == A.n and C.n == A.m), and C must have enough storage to accommodate all allocated entries in A (length(C.rowval) >= nnz(A)andlength(C.nzval) >= nnz(A)`).

For additional (algorithmic) information, and for versions of these methods that forgo argument checking, see (unexported) parent methods unchecked_noalias_permute! and unchecked_aliasing_permute!.

See also: permute

source

permute!(v, p)

Permute vector v in-place, according to permutation p. No checking is done to verify that p is a permutation.

To return a new permutation, use v[p]. Note that this is generally faster than permute!(v,p) for large vectors.

source

# Base.ipermute! — Function.

ipermute!(v, p)

Like permute!, but the inverse of the given permutation is applied.

source

# Base.Random.randcycle — Function.

randcycle([rng,] n)

Construct a random cyclic permutation of length n. The optional rng argument specifies a random number generator, see Random Numbers.

source

# Base.Random.shuffle — Function.

shuffle([rng,] v)

Return a randomly permuted copy of v. The optional rng argument specifies a random number generator (see Random Numbers). To permute v in-place, see shuffle!. To obtain randomly permuted indices, see randperm.

source

# Base.Random.shuffle! — Function.

shuffle!([rng,] v)

In-place version of shuffle: randomly permute the array v in-place, optionally supplying the random-number generator rng.

source

# Base.reverse — Function.

reverse(v [, start=1 [, stop=length(v) ]] )

Return a copy of v reversed from start to stop.

source

# Base.reverseind — Function.

reverseind(v, i)

Given an index i in reverse(v), return the corresponding index in v so that v[reverseind(v,i)] == reverse(v)[i]. (This can be nontrivial in the case where v is a Unicode string.)

source

# Base.reverse! — Function.

reverse!(v [, start=1 [, stop=length(v) ]]) -> v

In-place version of reverse.

source

BitArrays

BitArrays are space-efficient "packed" boolean arrays, which store one bit per boolean value. They can be used similarly to Array{Bool} arrays (which store one byte per boolean value), and can be converted to/from the latter via Array(bitarray) and BitArray(array), respectively.

Base.flipbits!
Base.rol!
Base.rol!
Base.rol
Base.ror!
Base.ror!
Base.ror

Sparse Vectors and Matrices

Sparse vectors and matrices largely support the same set of operations as their dense counterparts. The following functions are specific to sparse arrays.

Base.SparseArrays.sparse
Base.SparseArrays.sparsevec
Base.SparseArrays.sparsevec
Base.SparseArrays.issparse
Base.SparseArrays.sparse
Base.SparseArrays.sparsevec
Base.full
Base.SparseArrays.nnz
Base.SparseArrays.spzeros
Base.SparseArrays.spones
Base.SparseArrays.speye
Base.SparseArrays.speye
Base.SparseArrays.spdiagm
Base.SparseArrays.sprand
Base.SparseArrays.sprandn
Base.SparseArrays.nonzeros
Base.SparseArrays.rowvals
Base.SparseArrays.nzrange
Base.SparseArrays.dropzeros!
Base.SparseArrays.dropzeros
Base.SparseArrays.dropzeros!
Base.SparseArrays.dropzeros
Base.SparseArrays.permute
Base.permute!