Base.Markdown.Header{1}

Essentials

Base.Markdown.Header{2}

Introduction

Base.Markdown.Paragraph

The Julia standard library contains a range of functions and macros appropriate for performing scientific and numerical computing, but is also as broad as those of many general purpose programming languages. Additional functionality is available from a growing collection of available packages. Functions are grouped by topic below.

Base.Markdown.Paragraph

Some general notes:

Base.Markdown.List

Except for functions in built-in modules (Pkg, Collections, Test and Profile), all functions documented here are directly available for use in programs.

To use module functions, use import Module to import the module, and Module.fn(x) to use the functions.

Alternatively, using Module will import all exported Module functions into the current namespace.

By convention, function names ending with an exclamation point (!) modify their arguments. Some functions have both modifying (e.g., sort!) and non-modifying (sort) versions.

Base.Markdown.Header{2}

Getting Around

Base.Markdown.Code

Base.exit
Base.quit
Base.atexit
Base.atreplinit
Base.isinteractive
Base.whos
Base.summarysize
Base.edit
Base.edit
Base.@edit
Base.less
Base.less
Base.@less
Base.clipboard
Base.clipboard
Base.reload
Base.require
Base.compilecache
Base.__precompile__
Base.include_from_node1
Base.include_string
Base.include_dependency
Base.Docs.apropos
Base.which
Base.which
Base.@which
Base.methods
Base.methodswith
Base.@show
Base.versioninfo
Base.workspace
ans

Base.Markdown.Header{2}

All Objects

# Core.is — Function.
MethodsThis function has no methods.

# Core.isa — Function.
MethodsThis function has no methods.

# Base.isequal — Function.
MethodsThis function has no methods to display.Hiding 14 methods defined outside of this package.

# Base.isless — Function.
Methods

isless(
    x::Documenter.Formats.Format,
    y::Documenter.Formats.Format
)

defined at Enums.jl:84

Hiding 48 methods defined outside of this package.

# Base.ifelse — Function.
MethodsThis function has no methods to display.Hiding 5 methods defined outside of this package.

# Base.lexcmp — Function.
MethodsThis function has no methods to display.Hiding 5 methods defined outside of this package.

# Base.lexless — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Core.typeof — Function.
MethodsThis function has no methods.

# Core.tuple — Function.
MethodsThis function has no methods.

# Base.ntuple — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.object_id — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.hash — Function.
MethodsThis function has no methods to display.Hiding 45 methods defined outside of this package.

# Base.finalizer — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.finalize — Function.
MethodsThis function has no methods to display.Hiding 17 methods defined outside of this package.

# Base.copy — Function.
MethodsThis function has no methods to display.Hiding 38 methods defined outside of this package.

# Base.deepcopy — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Core.isdefined — Function.
MethodsThis function has no methods.

# Base.convert — Function.
Methods

convert(
    ::Type{Documenter.Formats.Format},
    x::Integer
)

defined at Enums.jl:79

Hiding 593 methods defined outside of this package.

# Base.promote — Function.
MethodsThis function has no methods to display.Hiding 5 methods defined outside of this package.

# Base.oftype — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.widen — Function.
MethodsThis function has no methods to display.Hiding 16 methods defined outside of this package.

# Base.identity — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

Base.Markdown.Header{2}

Types

Base.Markdown.Code

Base.supertype
Core.issubtype
Base.:<:
Base.subtypes
Base.typemin
Base.typemax
Base.realmin
Base.realmax
Base.maxintfloat
Base.sizeof
Base.eps
Base.eps
Base.eps
Base.promote_type
Base.promote_rule
Core.getfield
Core.setfield!
Base.fieldoffset
Core.fieldtype
Base.isimmutable
Base.isbits
Base.isleaftype
Base.typejoin
Base.typeintersect
Base.Val
Base.Enums.@enum
Base.instances

Base.Markdown.Header{2}

Generic Functions

# Base.method_exists — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Core.applicable — Function.
MethodsThis function has no methods.

# Core.invoke — Function.
MethodsThis function has no methods.

# Base.|> — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

Base.Markdown.Header{2}

Syntax

Base.Markdown.Code

Base.eval
Base.@eval
Base.evalfile
Base.esc
Base.gensym
Base.@gensym
Base.parse
Base.parse

Base.Markdown.Header{2}

Nullables

Base.Markdown.Code

Base.Nullable
Base.get
Base.isnull

Base.Markdown.Header{2}

System

Base.Markdown.Code

Base.run
Base.spawn
Base.DevNull
Base.success
Base.process_running
Base.process_exited
Base.kill
Base.Sys.set_process_title
Base.Sys.get_process_title
Base.readandwrite
Base.ignorestatus
Base.detach
Base.Cmd
Base.setenv
Base.withenv
Base.pipeline
Base.pipeline
Base.Libc.gethostname
Base.getipaddr
Base.Libc.getpid
Base.Libc.time
Base.time_ns
Base.tic
Base.toc
Base.toq
Base.@time
Base.@timev
Base.@timed
Base.@elapsed
Base.@allocated
Base.EnvHash
Base.ENV
Base.is_unix
Base.is_apple
Base.is_linux
Base.is_bsd
Base.is_windows
Base.Sys.windows_version
Base.@static

Base.Markdown.Header{2}

Errors

# Base.error — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Core.throw — Function.
MethodsThis function has no methods.

# Base.rethrow — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.backtrace — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.catch_backtrace — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.assert — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.@assert — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.ArgumentError — Type.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.AssertionError — Type.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Core.BoundsError — Type.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Base.DimensionMismatch — Type.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Core.DivideError — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Core.DomainError — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.EOFError — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Core.ErrorException — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Core.InexactError — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Core.InterruptException — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.KeyError — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.LoadError — Type.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.MethodError — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.NullException — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Core.OutOfMemoryError — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Core.ReadOnlyMemoryError — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Core.OverflowError — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.ParseError — Type.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.ProcessExitedException — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Core.StackOverflowError — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.SystemError — Type.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Core.TypeError — Type.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Core.UndefRefError — Type.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Core.UndefVarError — Type.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.InitError — Type.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.retry — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

Base.Markdown.Header{2}

Events

Base.Markdown.Code

Base.Timer
Base.Timer
Base.AsyncCondition
Base.AsyncCondition

Base.Markdown.Header{2}

Reflection

Base.Markdown.Code

Base.module_name
Base.module_parent
Base.current_module
Base.fullname
Base.names
Core.nfields
Base.fieldnames
Base.fieldname
Base.datatype_module
Base.isconst
Base.function_name
Base.function_module
Base.function_module
Base.functionloc
Base.functionloc
Base.@functionloc

Base.Markdown.Header{2}

Internals

# Base.gc — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.gc_enable — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.macroexpand — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.expand — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.code_lowered — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.@code_lowered — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.code_typed — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.@code_typed — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.code_warntype — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.@code_warntype — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.code_llvm — Function.
MethodsThis function has no methods to display.Hiding 6 methods defined outside of this package.

# Base.@code_llvm — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.code_native — Function.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Base.@code_native — Macro.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.precompile — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.