Collections and Data Structures

Iteration

Sequential iteration is implemented by the methods start(), done(), and next(). The general for loop:

for i = I   # or  "for i in I"
    # body
end

is translated into:

state = start(I)
while !done(I, state)
    (i, state) = next(I, state)
    # body
end

The state object may be anything, and should be chosen appropriately for each iterable type. See the manual section on the iteration interface for more details about defining a custom iterable type.

# Base.start — Function.

start(iter) -> state

Get initial iteration state for an iterable object.

source

# Base.done — Function.

done(iter, state) -> Bool

Test whether we are done iterating.

source

# Base.next — Function.

next(iter, state) -> item, state

For a given iterable object and iteration state, return the current item and the next iteration state.

source

# Base.zip — Function.

zip(iters...)

For a set of iterable objects, returns an iterable of tuples, where the ith tuple contains the ith component of each input iterable.

Note that zip is its own inverse: collect(zip(zip(a...)...)) == collect(a).

source

# Base.enumerate — Function.

enumerate(iter)

An iterator that yields (i, x) where i is an index starting at 1, and x is the ith value from the given iterator. It's useful when you need not only the values x over which you are iterating, but also the index i of the iterations.

julia> a = ["a", "b", "c"];

julia> for (index, value) in enumerate(a)
           println("$index $value")
       end
1 a
2 b
3 c

source

# Base.rest — Function.

rest(iter, state)

An iterator that yields the same elements as iter, but starting at the given state.

source

# Base.countfrom — Function.

countfrom(start=1, step=1)

An iterator that counts forever, starting at start and incrementing by step.

source

# Base.take — Function.

take(iter, n)

An iterator that generates at most the first n elements of iter.

source

# Base.drop — Function.

drop(iter, n)

An iterator that generates all but the first n elements of iter.

source

# Base.cycle — Function.

cycle(iter)

An iterator that cycles through iter forever.

source

# Base.repeated — Function.

repeated(x[, n::Int])

An iterator that generates the value x forever. If n is specified, generates x that many times (equivalent to take(repeated(x), n)).

source

# Base.iteratorsize — Function.

iteratorsize(itertype::Type) -> IteratorSize

Given the type of an iterator, returns one of the following values:

SizeUnknown() if the length (number of elements) cannot be determined in advance.

HasLength() if there is a fixed, finite length.

HasShape() if there is a known length plus a notion of multidimensional shape (as for an array). In this case the size function is valid for the iterator.

IsInfinite() if the iterator yields values forever.

The default value (for iterators that do not define this function) is HasLength(). This means that most iterators are assumed to implement length.

This trait is generally used to select between algorithms that pre-allocate space for their result, and algorithms that resize their result incrementally.

source

# Base.iteratoreltype — Function.

iteratoreltype(itertype::Type) -> IteratorEltype

Given the type of an iterator, returns one of the following values:

EltypeUnknown() if the type of elements yielded by the iterator is not known in advance.

HasEltype() if the element type is known, and eltype would return a meaningful value.

HasEltype() is the default, since iterators are assumed to implement eltype.

This trait is generally used to select between algorithms that pre-allocate a specific type of result, and algorithms that pick a result type based on the types of yielded values.

source

Fully implemented by:

Range

UnitRange

Tuple

Number

AbstractArray

IntSet

ObjectIdDict

Dict

WeakKeyDict

EachLine

AbstractString

Set

Task

General Collections

Base.isempty
Base.empty!
Base.length
Base.endof

Fully implemented by:

Range

UnitRange

Tuple

Number

AbstractArray

IntSet

Dict

WeakKeyDict

AbstractString

Set

Iterable Collections

Base.in
Base.eltype
Base.indexin
Base.findin
Base.unique
Base.unique
Base.unique
Base.allunique
Base.reduce
Base.reduce
Base.foldl
Base.foldl
Base.foldr
Base.foldr
Base.maximum
Base.maximum
Base.maximum!
Base.minimum
Base.minimum
Base.minimum!
Base.extrema
Base.extrema
Base.indmax
Base.indmin
Base.findmax
Base.findmax
Base.findmin
Base.findmin
Base.findmax!
Base.findmin!
Base.maxabs
Base.maxabs
Base.maxabs!
Base.minabs
Base.minabs
Base.minabs!
Base.sum
Base.sum
Base.sum!
Base.sum
Base.sumabs
Base.sumabs
Base.sumabs!
Base.sumabs2
Base.sumabs2
Base.sumabs2!
Base.prod
Base.prod
Base.prod!
Base.any
Base.any
Base.any!
Base.all
Base.all
Base.all!
Base.count
Base.any
Base.all
Base.foreach
Base.map
Base.map!
Base.map!
Base.mapreduce
Base.mapreduce
Base.mapfoldl
Base.mapfoldl
Base.mapfoldr
Base.mapfoldr
Base.first
Base.last
Base.step
Base.collect
Base.collect
Base.issubset
Base.filter
Base.filter!

Indexable Collections

Base.getindex
Base.setindex!

Fully implemented by:

Array

BitArray

AbstractArray

SubArray

ObjectIdDict

Dict

WeakKeyDict

AbstractString

Partially implemented by:

Range

UnitRange

Tuple

Associative Collections

Dict is the standard associative collection. Its implementation uses hash() as the hashing function for the key, and isequal() to determine equality. Define these two functions for custom types to override how they are stored in a hash table.

ObjectIdDict is a special hash table where the keys are always object identities.

WeakKeyDict is a hash table implementation where the keys are weak references to objects, and thus may be garbage collected even when referenced in a hash table.

Dicts can be created by passing pair objects constructed with =>() to a Dict constructor: Dict("A"=>1, "B"=>2). This call will attempt to infer type information from the keys and values (i.e. this example creates a Dict{String, Int64}). To explicitly specify types use the syntax Dict{KeyType,ValueType}(...). For example, Dict{String,Int32}("A"=>1, "B"=>2).

Dicts may also be created with generators. For example, Dict(i => f(i) for i = 1:10).

Given a dictionary D, the syntax D[x] returns the value of key x (if it exists) or throws an error, and D[x] = y stores the key-value pair x => y in D (replacing any existing value for the key x). Multiple arguments to D[...] are converted to tuples; for example, the syntax D[x,y] is equivalent to D[(x,y)], i.e. it refers to the value keyed by the tuple (x,y).

Base.Dict
Base.haskey
Base.get
Base.get
Base.get!
Base.get!
Base.getkey
Base.delete!
Base.pop!
Base.keys
Base.values
Base.merge
Base.merge!
Base.sizehint!
Base.keytype
Base.valtype

Fully implemented by:

ObjectIdDict

Dict

WeakKeyDict

Partially implemented by:

IntSet

Set

EnvHash

Array

BitArray

Set-Like Collections

Base.Set
Base.IntSet
Base.union
Base.union!
Base.intersect
Base.setdiff
Base.setdiff!
Base.symdiff
Base.symdiff!
Base.symdiff!
Base.symdiff!
Base.intersect!
Base.issubset

Fully implemented by:

IntSet

Set

Partially implemented by:

Array

Dequeues

Base.push!
Base.pop!
Base.unshift!
Base.shift!
Base.insert!
Base.deleteat!
Base.deleteat!
Base.splice!
Base.splice!
Base.resize!
Base.append!
Base.prepend!

Fully implemented by:

Vector (a.k.a. 1-dimensional Array)

BitVector (a.k.a. 1-dimensional BitArray)

PriorityQueue

The PriorityQueue type is available from the Collections module. It provides a basic priority queue implementation allowing for arbitrary key and priority types. Multiple identical keys are not permitted, but the priority of existing keys can be changed efficiently.

# Base.Collections.PriorityQueue — Type.

PriorityQueue(K, V, [ord])

Construct a new PriorityQueue, with keys of type K and values/priorites of type V. If an order is not given, the priority queue is min-ordered using the default comparison for V.

A PriorityQueue acts like a Dict, mapping values to their priorities, with the addition of a dequeue! function to remove the lowest priority element.

source

# Base.Collections.enqueue! — Function.

enqueue!(pq, k, v)

Insert the a key k into a priority queue pq with priority v.

source

# Base.Collections.dequeue! — Function.

dequeue!(pq)

Remove and return the lowest priority key from a priority queue.

source

# Base.Collections.peek — Function.

peek(pq)

Return the lowest priority key from a priority queue without removing that key from the queue.

source

PriorityQueue also behaves similarly to a Dict in that keys can be inserted and priorities accessed or changed using indexing notation.

julia> # Julia code
       pq = Collections.PriorityQueue();

julia> # Insert keys with associated priorities
       pq["a"] = 10; pq["b"] = 5; pq["c"] = 15; pq
Base.Collections.PriorityQueue{Any,Any,Base.Order.ForwardOrdering} with 3 entries:
  "c" => 15
  "b" => 5
  "a" => 10

julia> # Change the priority of an existing key
       pq["a"] = 0; pq
Base.Collections.PriorityQueue{Any,Any,Base.Order.ForwardOrdering} with 3 entries:
  "c" => 15
  "b" => 5
  "a" => 0

Heap Functions

Along with the PriorityQueue type, the Collections module provides lower level functions for performing binary heap operations on arrays. Each function takes an optional ordering argument. If not given, default ordering is used, so that elements popped from the heap are given in ascending order.

# Base.Collections.heapify — Function.

heapify(v, [ord])

Returns a new vector in binary heap order, optionally using the given ordering.

source

# Base.Collections.heapify! — Function.

heapify!(v, [ord])

In-place heapify.

source

# Base.Collections.isheap — Function.

isheap(v, [ord])

Return true if an array is heap-ordered according to the given order.

source

# Base.Collections.heappush! — Function.

heappush!(v, x, [ord])

Given a binary heap-ordered array, push a new element x, preserving the heap property. For efficiency, this function does not check that the array is indeed heap-ordered.

source

# Base.Collections.heappop! — Function.

heappop!(v, [ord])

Given a binary heap-ordered array, remove and return the lowest ordered element. For efficiency, this function does not check that the array is indeed heap-ordered.

source