Dates and Time

Dates and Time Types

Base.Dates.Period — Type.

Period
Year
Month
Week
Day
Hour
Minute
Second
Millisecond

Period types represent discrete, human representations of time.

source

Base.Dates.CompoundPeriod — Type.

CompoundPeriod

A CompoundPeriod is useful for expressing time periods that are not a fixed multiple of smaller periods. For example, "a year and a day" is not a fixed number of days, but can be expressed using a CompoundPeriod. In fact, a CompoundPeriod is automatically generated by addition of different period types, e.g. Year(1) + Day(1) produces a CompoundPeriod result.

source

Base.Dates.Instant — Type.

Instant

Instant types represent integer-based, machine representations of time as continuous timelines starting from an epoch.

source

Base.Dates.UTInstant — Type.

UTInstant{T}

The UTInstant represents a machine timeline based on UT time (1 day = one revolution of the earth). The T is a Period parameter that indicates the resolution or precision of the instant.

source

Base.Dates.TimeType — Type.

TimeType

TimeType types wrap Instant machine instances to provide human representations of the machine instant. Both DateTime and Date are subtypes of TimeType.

source

Base.Dates.DateTime — Type.

DateTime

DateTime wraps a UTInstant{Millisecond} and interprets it according to the proleptic Gregorian calendar.

source

Base.Dates.Date — Type.

Date

Date wraps a UTInstant{Day} and interprets it according to the proleptic Gregorian calendar.

source

Dates Functions

All Dates functions are defined in the Dates module; note that only the Date, DateTime, and now functions are exported; to use all other Dates functions, you'll need to prefix each function call with an explicit Dates., e.g. Dates.dayofweek(dt). Alternatively, you can write using Base.Dates to bring all exported functions into Main to be used without the Dates. prefix.

Base.Dates.DateTime
Base.Dates.DateTime
Base.Dates.DateTime
Base.Dates.DateTime
Base.Dates.DateTime
Base.Dates.format
Base.Dates.DateFormat
Base.Dates.DateTime
Base.Dates.Date
Base.Dates.Date
Base.Dates.Date
Base.Dates.Date
Base.Dates.Date
Base.Dates.Date
Base.Dates.now
Base.Dates.now
Base.eps

Accessor Functions

Base.Dates.year — Function.

year(dt::TimeType) -> Int64

The year of a Date or DateTime as an Int64.

source

Base.Dates.month — Function.

month(dt::TimeType) -> Int64

The month of a Date or DateTime as an Int64.

source

Base.Dates.week — Function.

week(dt::TimeType) -> Int64

Return the ISO week date of a Date or DateTime as an Int64. Note that the first week of a year is the week that contains the first Thursday of the year which can result in dates prior to January 4th being in the last week of the previous year. For example week(Date(2005,1,1)) is the 53rd week of 2004.

source

Base.Dates.day — Function.

day(dt::TimeType) -> Int64

The day of month of a Date or DateTime as an Int64.

source

Base.Dates.hour — Function.

hour(dt::DateTime) -> Int64

The hour of day of a DateTime as an Int64.

source

Base.Dates.minute — Function.

minute(dt::DateTime) -> Int64

The minute of a DateTime as an Int64.

source

Base.Dates.second — Function.

second(dt::DateTime) -> Int64

The second of a DateTime as an Int64.

source

Base.Dates.millisecond — Function.

millisecond(dt::DateTime) -> Int64

The millisecond of a DateTime as an Int64.

source

Base.Dates.Year — Type.

Year(dt::TimeType) -> Year

The year part of a Date or DateTime as a Year.

source

Year(v)

Construct a Year object with the given v value. Input must be losslessly convertible to an Int64.

source

Base.Dates.Month — Type.

Month(dt::TimeType) -> Month

The month part of a Date or DateTime as a Month.

source

Month(v)

Construct a Month object with the given v value. Input must be losslessly convertible to an Int64.

source

Base.Dates.Week — Type.

Week(dt::TimeType) -> Week

The week part of a Date or DateTime as a Week. For details see week(::TimeType).

source

Week(v)

Construct a Week object with the given v value. Input must be losslessly convertible to an Int64.

source

Base.Dates.Day — Type.

Day(dt::TimeType) -> Day

The day part of a Date or DateTime as a Day.

source

Day(v)

Construct a Day object with the given v value. Input must be losslessly convertible to an Int64.

source

Base.Dates.Hour — Type.

Hour(dt::DateTime) -> Hour

The hour part of a DateTime as a Hour.

source

Hour(v)

Construct a Hour object with the given v value. Input must be losslessly convertible to an Int64.

source

Base.Dates.Minute — Type.

Minute(dt::DateTime) -> Minute

The minute part of a DateTime as a Minute.

source

Minute(v)

Construct a Minute object with the given v value. Input must be losslessly convertible to an Int64.

source

Base.Dates.Second — Type.

Second(dt::DateTime) -> Second

The second part of a DateTime as a Second.

source

Second(v)

Construct a Second object with the given v value. Input must be losslessly convertible to an Int64.

source

Base.Dates.Millisecond — Type.

Millisecond(dt::DateTime) -> Millisecond

The millisecond part of a DateTime as a Millisecond.

source

Millisecond(v)

Construct a Millisecond object with the given v value. Input must be losslessly convertible to an Int64.

source

Base.Dates.yearmonth — Function.

yearmonth(dt::TimeType) -> (Int64, Int64)

Simultaneously return the year and month parts of a Date or DateTime.

source

Base.Dates.monthday — Function.

monthday(dt::TimeType) -> (Int64, Int64)

Simultaneously return the month and day parts of a Date or DateTime.

source

Base.Dates.yearmonthday — Function.

yearmonthday(dt::TimeType) -> (Int64, Int64, Int64)

Simultaneously return the year, month and day parts of a Date or DateTime.

source

Query Functions

Base.Dates.dayname — Function.

dayname(dt::TimeType; locale="english") -> AbstractString

Return the full day name corresponding to the day of the week of the Date or DateTime in the given locale.

source

Base.Dates.dayabbr — Function.

dayabbr(dt::TimeType; locale="english") -> AbstractString

Return the abbreviated name corresponding to the day of the week of the Date or DateTime in the given locale.

source

Base.Dates.dayofweek — Function.

dayofweek(dt::TimeType) -> Int64

Returns the day of the week as an Int64 with 1 = Monday, 2 = Tuesday, etc..

source

Base.Dates.dayofmonth — Function.

dayofmonth(dt::TimeType) -> Int64

The day of month of a Date or DateTime as an Int64.

source

Base.Dates.dayofweekofmonth — Function.

dayofweekofmonth(dt::TimeType) -> Int

For the day of week of dt, returns which number it is in dt's month. So if the day of the week of dt is Monday, then 1 = First Monday of the month, 2 = Second Monday of the month, etc. In the range 1:5.

source

Base.Dates.daysofweekinmonth — Function.

daysofweekinmonth(dt::TimeType) -> Int

For the day of week of dt, returns the total number of that day of the week in dt's month. Returns 4 or 5. Useful in temporal expressions for specifying the last day of a week in a month by including dayofweekofmonth(dt) == daysofweekinmonth(dt) in the adjuster function.

source

Base.Dates.monthname — Function.

monthname(dt::TimeType; locale="english") -> AbstractString

Return the full name of the month of the Date or DateTime in the given locale.

source

Base.Dates.monthabbr — Function.

monthabbr(dt::TimeType; locale="english") -> AbstractString

Return the abbreviated month name of the Date or DateTime in the given locale.

source

Base.Dates.daysinmonth — Function.

daysinmonth(dt::TimeType) -> Int

Returns the number of days in the month of dt. Value will be 28, 29, 30, or 31.

source

Base.Dates.isleapyear — Function.

isleapyear(dt::TimeType) -> Bool

Returns true if the year of dt is a leap year.

source

Base.Dates.dayofyear — Function.

dayofyear(dt::TimeType) -> Int

Returns the day of the year for dt with January 1st being day 1.

source

Base.Dates.daysinyear — Function.

daysinyear(dt::TimeType) -> Int

Returns 366 if the year of dt is a leap year, otherwise returns 365.

source

Base.Dates.quarterofyear — Function.

quarterofyear(dt::TimeType) -> Int

Returns the quarter that dt resides in. Range of value is 1:4.

source

Base.Dates.dayofquarter — Function.

dayofquarter(dt::TimeType) -> Int

Returns the day of the current quarter of dt. Range of value is 1:92.

source

Adjuster Functions

Base.trunc
Base.Dates.firstdayofweek
Base.Dates.lastdayofweek
Base.Dates.firstdayofmonth
Base.Dates.lastdayofmonth
Base.Dates.firstdayofyear
Base.Dates.lastdayofyear
Base.Dates.firstdayofquarter
Base.Dates.lastdayofquarter
Base.Dates.tonext
Base.Dates.toprev
Base.Dates.tofirst
Base.Dates.tolast
Base.Dates.tonext
Base.Dates.toprev
Base.Dates.recur

Periods

Base.Dates.Period
Base.Dates.CompoundPeriod
Base.Dates.default

Rounding Functions

Date and DateTime values can be rounded to a specified resolution (e.g., 1 month or 15 minutes) with floor, ceil, or round.

Base.floor — Function.

floor([T,] x, [digits, [base]])

floor(x) returns the nearest integral value of the same type as x that is less than or equal to x.

floor(T, x) converts the result to type T, throwing an InexactError if the value is not representable.

digits and base work as for round.

source

Base.ceil — Function.

ceil([T,] x, [digits, [base]])

ceil(x) returns the nearest integral value of the same type as x that is greater than or equal to x.

ceil(T, x) converts the result to type T, throwing an InexactError if the value is not representable.

digits and base work as for round.

source

Base.round — Function.

round(dt::TimeType, p::Period, [r::RoundingMode]) -> TimeType

Returns the Date or DateTime nearest to dt at resolution p. By default (RoundNearestTiesUp), ties (e.g., rounding 9:30 to the nearest hour) will be rounded up.

For convenience, p may be a type instead of a value: round(dt, Dates.Hour) is a shortcut for round(dt, Dates.Hour(1)).

julia> round(Date(1985, 8, 16), Dates.Month)
1985-08-01

julia> round(DateTime(2013, 2, 13, 0, 31, 20), Dates.Minute(15))
2013-02-13T00:30:00

julia> round(DateTime(2016, 8, 6, 12, 0, 0), Dates.Day)
2016-08-07T00:00:00

Valid rounding modes for round(::TimeType, ::Period, ::RoundingMode) are RoundNearestTiesUp (default), RoundDown (floor), and RoundUp (ceil).

source

round(z, RoundingModeReal, RoundingModeImaginary)

Returns the nearest integral value of the same type as the complex-valued z to z, breaking ties using the specified RoundingModes. The first RoundingMode is used for rounding the real components while the second is used for rounding the imaginary components.

source

round([T,] x, [digits, [base]], [r::RoundingMode])

Rounds x to an integer value according to the provided RoundingMode, returning a value of the same type as x. When not specifying a rounding mode the global mode will be used (see rounding), which by default is round to the nearest integer (RoundNearest mode), with ties (fractional values of 0.5) being rounded to the nearest even integer.

julia> round(1.7)
2.0

julia> round(1.5)
2.0

julia> round(2.5)
2.0

The optional RoundingMode argument will change how the number gets rounded.

round(T, x, [r::RoundingMode]) converts the result to type T, throwing an InexactError if the value is not representable.

round(x, digits) rounds to the specified number of digits after the decimal place (or before if negative). round(x, digits, base) rounds using a base other than 10.

julia> round(pi, 2)
3.14

julia> round(pi, 3, 2)
3.125

Note

Rounding to specified digits in bases other than 2 can be inexact when operating on binary floating point numbers. For example, the Float64 value represented by 1.15 is actually less than 1.15, yet will be rounded to 1.2.
julia> x = 1.15 1.15 julia> @sprintf "%.20f" x "1.14999999999999991118" julia> x < 115//100 true julia> round(x, 1) 1.2

source

The following functions are not exported:

Base.Dates.floorceil — Function.

floorceil(dt::TimeType, p::Period) -> (TimeType, TimeType)

Simultaneously return the floor and ceil of a Date or DateTime at resolution p. More efficient than calling both floor and ceil individually.

source

Base.Dates.epochdays2date — Function.

epochdays2date(days) -> Date

Takes the number of days since the rounding epoch (0000-01-01T00:00:00) and returns the corresponding Date.

source

Base.Dates.epochms2datetime — Function.

epochms2datetime(milliseconds) -> DateTime

Takes the number of milliseconds since the rounding epoch (0000-01-01T00:00:00) and returns the corresponding DateTime.

source

Base.Dates.date2epochdays — Function.

date2epochdays(dt::Date) -> Int64

Takes the given Date and returns the number of days since the rounding epoch (0000-01-01T00:00:00) as an Int64.

source

Base.Dates.datetime2epochms — Function.

datetime2epochms(dt::DateTime) -> Int64

Takes the given DateTime and returns the number of milliseconds since the rounding epoch (0000-01-01T00:00:00) as an Int64.

source

Conversion Functions

Base.Dates.today — Function.

today() -> Date

Returns the date portion of now().

source

Base.Dates.unix2datetime — Function.

unix2datetime(x) -> DateTime

Takes the number of seconds since unix epoch 1970-01-01T00:00:00 and converts to the corresponding DateTime.

source

Base.Dates.datetime2unix — Function.

datetime2unix(dt::DateTime) -> Float64

Takes the given DateTime and returns the number of seconds since the unix epoch 1970-01-01T00:00:00 as a Float64.

source

Base.Dates.julian2datetime — Function.

julian2datetime(julian_days) -> DateTime

Takes the number of Julian calendar days since epoch -4713-11-24T12:00:00 and returns the corresponding DateTime.

source

Base.Dates.datetime2julian — Function.

datetime2julian(dt::DateTime) -> Float64

Takes the given DateTime and returns the number of Julian calendar days since the julian epoch -4713-11-24T12:00:00 as a Float64.

source

Base.Dates.rata2datetime — Function.

rata2datetime(days) -> DateTime

Takes the number of Rata Die days since epoch 0000-12-31T00:00:00 and returns the corresponding DateTime.

source

Base.Dates.datetime2rata — Function.

datetime2rata(dt::TimeType) -> Int64

Returns the number of Rata Die days since epoch from the given Date or DateTime.

source

Constants

Days of the Week:

Variable	Abbr.	Value (Int)

Monday	Mon	1

Tuesday	Tue	2

Wednesday	Wed	3

Thursday	Thu	4

Friday	Fri	5

Saturday	Sat	6

Sunday	Sun	7

Months of the Year:

Variable	Abbr.	Value (Int)

January	Jan	1

February	Feb	2

March	Mar	3

April	Apr	4

May	May	5

June	Jun	6

July	Jul	7

August	Aug	8

September	Sep	9

October	Oct	10

November	Nov	11

December	Dec	12