Base.Markdown.Header{1}

I/O and Network

Base.Markdown.Header{2}

General I/O

Base.Markdown.Code

Base.STDOUT
Base.STDERR
Base.STDIN
Base.open
Base.open
Base.open
Base.open
Base.open
Base.IOBuffer
Base.IOBuffer
Base.IOBuffer
Base.IOBuffer
Base.takebuf_array
Base.takebuf_string
Base.fdio
Base.flush
Base.close
Base.write
Base.read
Base.read
Base.read!
Base.readbytes!
Base.read
Base.read
Base.unsafe_read
Base.unsafe_write
Base.position
Base.seek
Base.seekstart
Base.seekend
Base.skip
Base.mark
Base.unmark
Base.reset
Base.ismarked
Base.eof
Base.isreadonly
Base.iswritable
Base.isreadable
Base.isopen
Base.Serializer.serialize
Base.Serializer.deserialize
Base.escape_string
Base.unescape_string
Base.join
Base.Grisu.print_shortest
Base.fd
Base.redirect_stdout
Base.redirect_stdout
Base.redirect_stderr
Base.redirect_stdin
Base.readchomp
Base.truncate
Base.skipchars
Base.DataFmt.countlines
Base.PipeBuffer
Base.PipeBuffer
Base.readavailable
Base.IOContext
Base.IOContext
Base.IOContext
Base.IOContext

Base.Markdown.Header{2}

Text I/O

Base.Markdown.Code

Base.show
Base.showcompact
Base.showall
Base.summary
Base.print
Base.println
Base.print_with_color
Base.info
Base.warn
Base.Printf.@printf
Base.Printf.@sprintf
Base.sprint
Base.showerror
Base.dump
Base.readstring
Base.readline
Base.readuntil
Base.readlines
Base.eachline
Base.DataFmt.readdlm
Base.DataFmt.readdlm
Base.DataFmt.readdlm
Base.DataFmt.readdlm
Base.DataFmt.readdlm
Base.DataFmt.readdlm
Base.DataFmt.writedlm
Base.DataFmt.readcsv
Base.DataFmt.writecsv
Base.Base64.Base64EncodePipe
Base.Base64.Base64DecodePipe
Base.Base64.base64encode
Base.Base64.base64decode
Base.displaysize

Base.Markdown.Header{2}

Multimedia I/O

Base.Markdown.Paragraph

Just as text output is performed by print and user-defined types can indicate their textual representation by overloading show, Julia provides a standardized mechanism for rich multimedia output (such as images, formatted text, or even audio and video), consisting of three parts:

Base.Markdown.List

A function display(x) to request the richest available multimedia display of a Julia object x (with a plain-text fallback).

Overloading show allows one to indicate arbitrary multimedia representations (keyed by standard MIME types) of user-defined types.

Multimedia-capable display backends may be registered by subclassing a generic Display type and pushing them onto a stack of display backends via pushdisplay.

Base.Markdown.Paragraph

The base Julia runtime provides only plain-text display, but richer displays may be enabled by loading external modules or by using graphical Julia environments (such as the IPython-based IJulia notebook).

Base.Markdown.Code

Base.Multimedia.display
Base.Multimedia.redisplay
Base.Multimedia.displayable
Base.show
Base.Multimedia.mimewritable
Base.Multimedia.reprmime
Base.Multimedia.stringmime

Base.Markdown.Paragraph

As mentioned above, one can also define new display backends. For example, a module that can display PNG images in a window can register this capability with Julia, so that calling display(x) on types with PNG representations will automatically display the image using the module's window.

Base.Markdown.Paragraph

In order to define a new display backend, one should first create a subtype D of the abstract class Display. Then, for each MIME type (mime string) that can be displayed on D, one should define a function display(d::D, ::MIME"mime", x) = ... that displays x as that MIME type, usually by calling reprmime(mime, x). A MethodError should be thrown if x cannot be displayed as that MIME type; this is automatic if one calls reprmime. Finally, one should define a function display(d::D, x) that queries mimewritable(mime, x) for the mime types supported by D and displays the "best" one; a MethodError should be thrown if no supported MIME types are found for x. Similarly, some subtypes may wish to override redisplay(d::D, ...). (Again, one should import Base.display to add new methods to display.) The return values of these functions are up to the implementation (since in some cases it may be useful to return a display "handle" of some type). The display functions for D can then be called directly, but they can also be invoked automatically from display(x) simply by pushing a new display onto the display-backend stack with:

# Base.Multimedia.pushdisplay — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.Multimedia.popdisplay — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Multimedia.TextDisplay — Type.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Multimedia.istextmime — Function.
MethodsThis function has no methods to display.Hiding 36 methods defined outside of this package.

Base.Markdown.Header{2}

Memory-mapped I/O

Base.Markdown.Code

Base.Mmap.Anonymous
Base.Mmap.mmap
Base.Mmap.mmap
Base.Mmap.sync!

Base.Markdown.Header{2}

Network I/O

Base.Markdown.Code

Base.connect
Base.connect
Base.listen
Base.listen
Base.getaddrinfo
Base.getsockname
Base.IPv4
Base.IPv6
Base.nb_available
Base.accept
Base.listenany
Base.Filesystem.poll_fd
Base.Filesystem.poll_file
Base.Filesystem.watch_file
Base.bind
Base.send
Base.recv
Base.recvfrom
Base.setopt
Base.ntoh
Base.hton
Base.ltoh
Base.htol
Base.ENDIAN_BOM