Base.Markdown.Header{1}

Numbers

Base.Markdown.Header{2}

Standard Numeric Types

Base.Markdown.List

Bool

Int8

UInt8

Int16

UInt16

Int32

UInt32

Int64

UInt64

Int128

UInt128

Float16

Float32

Float64

Complex64

Complex128

Base.Markdown.Header{2}

Data Formats

Base.Markdown.Code

Base.bin
Base.hex
Base.dec
Base.oct
Base.base
Base.digits
Base.digits!
Base.bits
Base.parse
Base.tryparse
Base.big
Base.signed
Base.unsigned
Base.float
Base.Math.significand
Base.Math.exponent
Base.complex
Base.bswap
Base.num2hex
Base.hex2num
Base.hex2bytes
Base.bytes2hex

Base.Markdown.Header{2}

General Number Functions and Constants

Base.Markdown.Code

Base.one
Base.zero
Base.pi
Base.im
Base.eu
Base.catalan
Base.eulergamma
Base.golden
Base.Inf
Base.Inf32
Base.Inf16
Base.NaN
Base.NaN32
Base.NaN16
Base.issubnormal
Base.isfinite
Base.isinf
Base.isnan
Base.nextfloat
Base.prevfloat
Base.nextfloat
Base.isinteger
Base.isreal
Base.isimag
Core.Float32
Core.Float64
Base.GMP.BigInt
Base.MPFR.BigFloat
Base.Rounding.rounding
Base.Rounding.setrounding
Base.Rounding.setrounding
Base.Rounding.get_zero_subnormals
Base.Rounding.set_zero_subnormals

Base.Markdown.Header{3}

Integers

# Base.count_ones — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.count_zeros — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.leading_zeros — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.leading_ones — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.trailing_zeros — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.trailing_ones — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.isodd — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

# Base.iseven — Function.
MethodsThis function has no methods to display.Hiding 1 method defined outside of this package.

Base.Markdown.Header{2}

BigFloats

Base.Markdown.Paragraph

The BigFloat type implements arbitrary-precision floating-point arithmetic using the GNU MPFR library.

Base.Markdown.Code

Base.precision
Base.precision
Base.MPFR.setprecision
Base.MPFR.setprecision

Base.Markdown.Header{2}

Random Numbers

Base.Markdown.Paragraph

Random number generation in Julia uses the Mersenne Twister library via MersenneTwister objects. Julia has a global RNG, which is used by default. Other RNG types can be plugged in by inheriting the AbstractRNG type; they can then be used to have multiple streams of random numbers. Besides MersenneTwister, Julia also provides the RandomDevice RNG type, which is a wrapper over the OS provided entropy.

Base.Markdown.Paragraph

Most functions related to random generation accept an optional AbstractRNG as the first argument, rng , which defaults to the global one if not provided. Morever, some of them accept optionally dimension specifications dims... (which can be given as a tuple) to generate arrays of random values.

Base.Markdown.Paragraph

A MersenneTwister or RandomDevice RNG can generate random numbers of the following types: Float16, Float32, Float64, Bool, Int8, UInt8, Int16, UInt16, Int32, UInt32, Int64, UInt64, Int128, UInt128, BigInt (or complex numbers of those types). Random floating point numbers are generated uniformly in ([0, 1)). As BigInt represents unbounded integers, the interval must be specified (e.g. rand(big(1:6))).

# Base.Random.srand — Function.
MethodsThis function has no methods to display.Hiding 9 methods defined outside of this package.

# Base.Random.MersenneTwister — Type.
MethodsThis function has no methods to display.Hiding 5 methods defined outside of this package.

# Base.Random.RandomDevice — Type.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Random.rand — Function.
MethodsThis function has no methods to display.Hiding 35 methods defined outside of this package.

# Base.Random.rand! — Function.
MethodsThis function has no methods to display.Hiding 19 methods defined outside of this package.

# Base.Random.bitrand — Function.
MethodsThis function has no methods to display.Hiding 4 methods defined outside of this package.

# Base.Random.randn — Function.
MethodsThis function has no methods to display.Hiding 6 methods defined outside of this package.

# Base.Random.randn! — Function.
MethodsThis function has no methods to display.Hiding 3 methods defined outside of this package.

# Base.Random.randexp — Function.
MethodsThis function has no methods to display.Hiding 6 methods defined outside of this package.

# Base.Random.randexp! — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.

# Base.Random.randjump — Function.
MethodsThis function has no methods to display.Hiding 2 methods defined outside of this package.