Numbers

Standard Numeric Types

Bool

Int8

UInt8

Int16

UInt16

Int32

UInt32

Int64

UInt64

Int128

UInt128

Float16

Float32

Float64

Complex64

Complex128

Data Formats

# Base.bin — Function.

bin(n, [pad])

Convert an integer to a binary string, optionally specifying a number of digits to pad to.

source

# Base.hex — Function.

hex(n, [pad])

Convert an integer to a hexadecimal string, optionally specifying a number of digits to pad to.

source

# Base.dec — Function.

dec(n, [pad])

Convert an integer to a decimal string, optionally specifying a number of digits to pad to.

source

# Base.oct — Function.

oct(n, [pad])

Convert an integer to an octal string, optionally specifying a number of digits to pad to.

source

# Base.base — Function.

base(base, n, [pad])

Convert an integer to a string in the given base, optionally specifying a number of digits to pad to.

source

# Base.digits — Function.

digits([T], n, [base], [pad])

Returns an array with element type T (default Int) of the digits of n in the given base, optionally padded with zeros to a specified size. More significant digits are at higher indexes, such that n == sum([digits[k]*base^(k-1) for k=1:length(digits)]).

source

# Base.digits! — Function.

digits!(array, n, [base])

Fills an array of the digits of n in the given base. More significant digits are at higher indexes. If the array length is insufficient, the least significant digits are filled up to the array length. If the array length is excessive, the excess portion is filled with zeros.

source

# Base.bits — Function.

bits(n)

A string giving the literal bit representation of a number.

source

# Base.parse — Function.

parse(str, start; greedy=true, raise=true)

Parse the expression string and return an expression (which could later be passed to eval for execution). start is the index of the first character to start parsing. If greedy is true (default), parse will try to consume as much input as it can; otherwise, it will stop as soon as it has parsed a valid expression. Incomplete but otherwise syntactically valid expressions will return Expr(:incomplete, "(error message)"). If raise is true (default), syntax errors other than incomplete expressions will raise an error. If raise is false, parse will return an expression that will raise an error upon evaluation.

source

parse(str; raise=true)

Parse the expression string greedily, returning a single expression. An error is thrown if there are additional characters after the first expression. If raise is true (default), syntax errors will raise an error; otherwise, parse will return an expression that will raise an error upon evaluation.

source

parse(type, str, [base])

Parse a string as a number. If the type is an integer type, then a base can be specified (the default is 10). If the type is a floating point type, the string is parsed as a decimal floating point number. If the string does not contain a valid number, an error is raised.

source

# Base.tryparse — Function.

tryparse(type, str, [base])

Like parse, but returns a Nullable of the requested type. The result will be null if the string does not contain a valid number.

source

# Base.big — Function.

big(x)

Convert a number to a maximum precision representation (typically BigInt or BigFloat). See BigFloat for information about some pitfalls with floating-point numbers.

source

# Base.signed — Function.

signed(x)

Convert a number to a signed integer. If the argument is unsigned, it is reinterpreted as signed without checking for overflow.

source

# Base.unsigned — Function.

unsigned(x) -> Unsigned

Convert a number to an unsigned integer. If the argument is signed, it is reinterpreted as unsigned without checking for negative values.

source

# Base.float — Function.

float(x)

Convert a number, array, or string to a AbstractFloat data type. For numeric data, the smallest suitable AbstractFloat type is used. Converts strings to Float64.

source

# Base.Math.significand — Function.

significand(x)

Extract the significand(s) (a.k.a. mantissa), in binary representation, of a floating-point number or array. If x is a non-zero finite number, then the result will be a number of the same type on the interval $[1,2)$. Otherwise x is returned.

julia> significand(15.2)/15.2
0.125

julia> significand(15.2)*8
15.2

source

# Base.Math.exponent — Function.

exponent(x) -> Int

Get the exponent of a normalized floating-point number.

source

# Base.complex — Function.

complex(r, [i])

Convert real numbers or arrays to complex. i defaults to zero.

source

# Base.bswap — Function.

bswap(n)

Byte-swap an integer.

source

# Base.num2hex — Function.

num2hex(f)

Get a hexadecimal string of the binary representation of a floating point number.

source

# Base.hex2num — Function.

hex2num(str)

Convert a hexadecimal string to the floating point number it represents.

source

# Base.hex2bytes — Function.

hex2bytes(s::AbstractString)

Convert an arbitrarily long hexadecimal string to its binary representation. Returns an Array{UInt8,1}, i.e. an array of bytes.

source

# Base.bytes2hex — Function.

bytes2hex(bin_arr::Array{UInt8, 1})

Convert an array of bytes to its hexadecimal representation. All characters are in lower-case. Returns a String.

source

General Number Functions and Constants

Base.one
Base.zero
Base.pi
Base.im
Base.eu
Base.catalan
Base.eulergamma
Base.golden
Base.Inf
Base.Inf32
Base.Inf16
Base.NaN
Base.NaN32
Base.NaN16
Base.issubnormal
Base.isfinite
Base.isinf
Base.isnan
Base.nextfloat
Base.prevfloat
Base.nextfloat
Base.isinteger
Base.isreal
Base.isimag
Core.Float32
Core.Float64
Base.GMP.BigInt
Base.MPFR.BigFloat
Base.Rounding.rounding
Base.Rounding.setrounding
Base.Rounding.setrounding
Base.Rounding.get_zero_subnormals
Base.Rounding.set_zero_subnormals

Integers

# Base.count_ones — Function.

count_ones(x::Integer) -> Integer

Number of ones in the binary representation of x.

julia> count_ones(7)
3

source

# Base.count_zeros — Function.

count_zeros(x::Integer) -> Integer

Number of zeros in the binary representation of x.

julia> count_zeros(Int32(2 ^ 16 - 1))
16

source

# Base.leading_zeros — Function.

leading_zeros(x::Integer) -> Integer

Number of zeros leading the binary representation of x.

julia> leading_zeros(Int32(1))
31

source

# Base.leading_ones — Function.

leading_ones(x::Integer) -> Integer

Number of ones leading the binary representation of x.

julia> leading_ones(UInt32(2 ^ 32 - 2))
31

source

# Base.trailing_zeros — Function.

trailing_zeros(x::Integer) -> Integer

Number of zeros trailing the binary representation of x.

julia> trailing_zeros(2)
1

source

# Base.trailing_ones — Function.

trailing_ones(x::Integer) -> Integer

Number of ones trailing the binary representation of x.

julia> trailing_ones(3)
2

source

# Base.isodd — Function.

isodd(x::Integer) -> Bool

Returns true if x is odd (that is, not divisible by 2), and false otherwise.

julia> isodd(9)
true

julia> isodd(10)
false

source

# Base.iseven — Function.

iseven(x::Integer) -> Bool

Returns true is x is even (that is, divisible by 2), and false otherwise.

julia> iseven(9)
false

julia> iseven(10)
true

source

BigFloats

The BigFloat type implements arbitrary-precision floating-point arithmetic using the GNU MPFR library.

Base.precision
Base.precision
Base.MPFR.setprecision
Base.MPFR.setprecision

Random Numbers

Random number generation in Julia uses the Mersenne Twister library via MersenneTwister objects. Julia has a global RNG, which is used by default. Other RNG types can be plugged in by inheriting the AbstractRNG type; they can then be used to have multiple streams of random numbers. Besides MersenneTwister, Julia also provides the RandomDevice RNG type, which is a wrapper over the OS provided entropy.

Most functions related to random generation accept an optional AbstractRNG as the first argument, rng , which defaults to the global one if not provided. Morever, some of them accept optionally dimension specifications dims... (which can be given as a tuple) to generate arrays of random values.

A MersenneTwister or RandomDevice RNG can generate random numbers of the following types: Float16, Float32, Float64, Bool, Int8, UInt8, Int16, UInt16, Int32, UInt32, Int64, UInt64, Int128, UInt128, BigInt (or complex numbers of those types). Random floating point numbers are generated uniformly in ([0, 1)). As BigInt represents unbounded integers, the interval must be specified (e.g. rand(big(1:6))).

# Base.Random.srand — Function.

srand([rng], [seed])

Reseed the random number generator. If a seed is provided, the RNG will give a reproducible sequence of numbers, otherwise Julia will get entropy from the system. For MersenneTwister, the seed may be a non-negative integer, a vector of UInt32 integers or a filename, in which case the seed is read from a file. RandomDevice does not support seeding.

source

# Base.Random.MersenneTwister — Type.

MersenneTwister([seed])

Create a MersenneTwister RNG object. Different RNG objects can have their own seeds, which may be useful for generating different streams of random numbers.

source

# Base.Random.RandomDevice — Type.

RandomDevice()

Create a RandomDevice RNG object. Two such objects will always generate different streams of random numbers.

source

# Base.Random.rand — Function.

rand([rng], [S], [dims...])

Pick a random element or array of random elements from the set of values specified by S; S can be

an indexable collection (for example 1:n or ['x','y','z']), or

a type: the set of values to pick from is then equivalent to typemin(S):typemax(S) for integers (this is not applicable to BigInt), and to $[0, 1)$ for floating point numbers;

S defaults to Float64.

source

# Base.Random.rand! — Function.

rand!([rng], A, [coll])

Populate the array A with random values. If the indexable collection coll is specified, the values are picked randomly from coll. This is equivalent to copy!(A, rand(rng, coll, size(A))) or copy!(A, rand(rng, eltype(A), size(A))) but without allocating a new array.

source

# Base.Random.bitrand — Function.

bitrand([rng], [dims...])

Generate a BitArray of random boolean values.

source

# Base.Random.randn — Function.

randn([rng], [T=Float64], [dims...])

Generate a normally-distributed random number of type T with mean 0 and standard deviation 1. Optionally generate an array of normally-distributed random numbers. The Base module currently provides an implementation for the types Float16, Float32, and Float64 (the default).

source

# Base.Random.randn! — Function.

randn!([rng], A::AbstractArray) -> A

Fill the array A with normally-distributed (mean 0, standard deviation 1) random numbers. Also see the rand function.

source

# Base.Random.randexp — Function.

randexp([rng], [T=Float64], [dims...])

Generate a random number of type T according to the exponential distribution with scale 1. Optionally generate an array of such random numbers. The Base module currently provides an implementation for the types Float16, Float32, and Float64 (the default).

source

# Base.Random.randexp! — Function.

randexp!([rng], A::AbstractArray) -> A

Fill the array A with random numbers following the exponential distribution (with scale 1).

source

# Base.Random.randjump — Function.

randjump(r::MersenneTwister, jumps, [jumppoly]) -> Vector{MersenneTwister}

Create an array of the size jumps of initialized MersenneTwister RNG objects where the first RNG object given as a parameter and following MersenneTwister RNGs in the array initialized such that a state of the RNG object in the array would be moved forward (without generating numbers) from a previous RNG object array element on a particular number of steps encoded by the jump polynomial jumppoly.

Default jump polynomial moves forward MersenneTwister RNG state by 10^20 steps.

source