Tasks and Parallel Computing

Tasks

Core.Task
Base.yieldto
Base.current_task
Base.istaskdone
Base.istaskstarted
Base.consume
Base.produce
Base.yield
Base.task_local_storage
Base.task_local_storage
Base.task_local_storage
Base.Condition
Base.notify
Base.schedule
Base.@schedule
Base.@task
Base.sleep
Base.ReentrantLock
Base.lock
Base.unlock
Base.Channel

General Parallel Computing Support

Base.addprocs
Base.addprocs
Base.addprocs
Base.addprocs
Base.nprocs
Base.nworkers
Base.procs
Base.workers
Base.rmprocs
Base.interrupt
Base.myid
Base.asyncmap
Base.pmap
Base.remotecall
Base.Future
Base.Future
Base.RemoteChannel
Base.RemoteChannel
Base.RemoteChannel
Base.wait
Base.fetch
Base.remotecall_wait
Base.remotecall_fetch
Base.put!
Base.put!
Base.put!
Base.take!
Base.take!
Base.isready
Base.isready
Base.close
Base.WorkerPool
Base.CachingPool
Base._default_worker_pool
Base.remote
Base.remotecall
Base.remotecall_wait
Base.remotecall_fetch
Base.timedwait
Base.@spawn
Base.@spawnat
Base.@fetch
Base.@fetchfrom
Base.@async
Base.@sync
Base.@parallel
Base.@everywhere
Base.clear!
Base.remoteref_id
Base.channel_from_id
Base.worker_id_from_socket
Base.cluster_cookie

Shared Arrays

Base.SharedArray
Base.SharedArray
Base.procs
Base.sdata
Base.indexpids
Base.localindexes

Multi-Threading

This experimental interface supports Julia's multi-threading capabilities. Types and function described here might (and likely will) change in the future.

Base.Threads.threadid — Function.

Threads.threadid()

Get the ID number of the current thread of execution. The master thread has ID 1.

source

Base.Threads.nthreads — Function.

Threads.nthreads()

Get the number of threads available to the Julia process. This is the inclusive upper bound on threadid().

source

Base.Threads.@threads — Macro.

Threads.@threads

A macro to parallelize a for-loop to run with multiple threads. This spawns nthreads() number of threads, splits the iteration space amongst them, and iterates in parallel. A barrier is placed at the end of the loop which waits for all the threads to finish execution, and the loop returns.

source

Base.Threads.Atomic — Type.

Threads.Atomic{T}

Holds a reference to an object of type T, ensuring that it is only accessed atomically, i.e. in a thread-safe manner.

Only certain "simple" types can be used atomically, namely the bitstypes integer and float-point types. These are Int8...Int128, UInt8...UInt128, and Float16...Float64.

New atomic objects can be created from a non-atomic values; if none is specified, the atomic object is initialized with zero.

Atomic objects can be accessed using the [] notation:

x::Atomic{Int}
x[] = 1
val = x[]

Atomic operations use an atomic_ prefix, such as atomic_add!, atomic_xchg!, etc.

source

Base.Threads.atomic_cas! — Function.

Threads.atomic_cas!{T}(x::Atomic{T}, cmp::T, newval::T)

Atomically compare-and-set x

Atomically compares the value in x with cmp. If equal, write newval to x. Otherwise, leaves x unmodified. Returns the old value in x. By comparing the returned value to cmp (via ===) one knows whether x was modified and now holds the new value newval.

For further details, see LLVM's cmpxchg instruction.

This function can be used to implement transactional semantics. Before the transaction, one records the value in x. After the transaction, the new value is stored only if x has not been modified in the mean time.

source

Base.Threads.atomic_xchg! — Function.

Threads.atomic_xchg!{T}(x::Atomic{T}, newval::T)

Atomically exchange the value in x

Atomically exchanges the value in x with newval. Returns the old value.

For further details, see LLVM's atomicrmw xchg instruction.

source

Base.Threads.atomic_add! — Function.

Threads.atomic_add!{T}(x::Atomic{T}, val::T)

Atomically add val to x

Performs x[] += val atomically. Returns the old (!) value.

For further details, see LLVM's atomicrmw add instruction.

source

Base.Threads.atomic_sub! — Function.

Threads.atomic_sub!{T}(x::Atomic{T}, val::T)

Atomically subtract val from x

Performs x[] -= val atomically. Returns the old (!) value.

For further details, see LLVM's atomicrmw sub instruction.

source

Base.Threads.atomic_and! — Function.

Threads.atomic_and!{T}(x::Atomic{T}, val::T)

Atomically bitwise-and x with val

Performs x[] &= val atomically. Returns the old (!) value.

For further details, see LLVM's atomicrmw and instruction.

source

Base.Threads.atomic_nand! — Function.

Threads.atomic_nand!{T}(x::Atomic{T}, val::T)

Atomically bitwise-nand (not-and) x with val

Performs x[] = ~(x[] & val) atomically. Returns the old (!) value.

For further details, see LLVM's atomicrmw nand instruction.

source

Base.Threads.atomic_or! — Function.

Threads.atomic_or!{T}(x::Atomic{T}, val::T)

Atomically bitwise-or x with val

Performs x[] |= val atomically. Returns the old (!) value.

For further details, see LLVM's atomicrmw or instruction.

source

Base.Threads.atomic_xor! — Function.

Threads.atomic_xor!{T}(x::Atomic{T}, val::T)

Atomically bitwise-xor (exclusive-or) x with val

Performs x[] $= val atomically. Returns the old (!) value.

For further details, see LLVM's atomicrmw xor instruction.

source

Base.Threads.atomic_max! — Function.

Threads.atomic_max!{T}(x::Atomic{T}, val::T)

Atomically store the maximum of x and val in x

Performs x[] = max(x[], val) atomically. Returns the old (!) value.

For further details, see LLVM's atomicrmw min instruction.

source

Base.Threads.atomic_min! — Function.

Threads.atomic_min!{T}(x::Atomic{T}, val::T)

Atomically store the minimum of x and val in x

Performs x[] = min(x[], val) atomically. Returns the old (!) value.

For further details, see LLVM's atomicrmw max instruction.

source

Base.Threads.atomic_fence — Function.

Threads.atomic_fence()

Insert a sequential-consistency memory fence

Inserts a memory fence with sequentially-consistent ordering semantics. There are algorithms where this is needed, i.e. where an acquire/release ordering is insufficient.

This is likely a very expensive operation. Given that all other atomic operations in Julia already have acquire/release semantics, explicit fences should not be necessary in most cases.

For further details, see LLVM's fence instruction.

source

Cluster Manager Interface

This interface provides a mechanism to launch and manage Julia workers on different cluster environments. LocalManager, for launching additional workers on the same host and SSHManager, for launching on remote hosts via ssh are present in Base. TCP/IP sockets are used to connect and transport messages between processes. It is possible for Cluster Managers to provide a different transport.

Base.launch
Base.manage
Base.kill
Base.init_worker
Base.connect
Base.process_messages